% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{add_pies}
\alias{add_pies}
\title{Add_pies on map or network for plot.data_agro, plot.data_network}
\usage{
add_pies(p, n, format, plot_type, data_to_pie, variable, pie_size)
}
\arguments{
\item{p}{network or map ggplot}

\item{n}{network object from igraph or data frame with coordinates}

\item{format}{format of n}

\item{plot_type}{network or map}

\item{data_to_pie}{data with the variables}

\item{variable}{variable to display}

\item{pie_size}{size of the pie}
}
\description{
Add_pies on map or network for plot.data_agro, plot.data_network
}
\details{
script adapted from 
Pies On A Map, Demonstration script, By QDR : 
https://qdrsite.wordpress.com/2016/06/26/pies-on-a-map/

Guangchuang YU code :
https://cran.r-project.org/web/packages/ggimage/vignettes/ggimage.html#geom_subview
https://github.com/GuangchuangYu/ggimage/blob/master/R/geom_subview.R
}
