% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{add_split_col}
\alias{add_split_col}
\title{add a column to split a dataframe}
\usage{
add_split_col(x, each)
}
\arguments{
\item{x}{vector}

\item{each}{nb of element iln each split}
}
\description{
add a column to split a dataframe
}
