% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AMMI.R
\name{AMMI}
\alias{AMMI}
\title{Run AMMI model on a data set}
\usage{
AMMI(data, vec_variables)
}
\arguments{
\item{data}{data with at least the following columns: year, location, germplasm, block, X, Y}

\item{vec_variables}{vectors of variables to analyse}
}
\value{
The function returns two lists with
\itemize{
 \item for each variable a list with
  \itemize{
   \item "model" being a list with
    \itemize{
     \item "anova_model"
     \item "residuals" being a list with
      \itemize{
       \item "distribution"
       \item "Residuals_vs_fitted"
       \item "QQplot"
       \item "standardized_residuals_vs_fitted"
      }
     \item "variability_repartition"
     \item "interaction_matrix"
     \item "location_effects"
     \item "germplasm_effects"
     \item "intra_germplasm_variance"
    }

   \item germplasm being a list with
     \itemize{
      \item "boxplot"
      \item "barplot_LSD_significant_group"
      \item "variance_intra"
     }

   \item location being a list with
    \itemize{
      \item "boxplot"
      \item "barplot_LSD_significant_group"
    }

   \item GxE being a list with
    \itemize{
     \item "interaction_plot"
     \item "ecovalence"
     \item "PCA" being a list with
      \itemize{
       \item "variation_dim"
       \item "CP1-CP2" being a list with
        \itemize{
         \item "ind"
         \item "var"
         }
       \item "CP2-CP3" being a list with
        \itemize{
         \item "ind"
         \item "var"
        }
      }
    }
  }

 \item a post AMMI analysis with
  \itemize{
   \item "barplot_variation_repartition"
   \item "PCA_G_effect"
   \item "PCA_intraG_effect"
   \item "PCA_E_effect"
  }
 }
}
\description{
\code{AMMI} runs AMMI model on a data set and gives back results on germplasms, locations, years and GxE interactions effects
}
\details{
See the vignette for more details on the AMMI model
}
\author{
Pierre Riviere
}

