% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyse.outputs.R
\name{analyse.outputs}
\alias{analyse.outputs}
\title{Check with plots if the model went well based on the Gelman-Rubin test and plots of posteriors distributions}
\usage{
analyse.outputs(out.model, analysis = NULL, nb_parameters_per_plot = 10)
}
\arguments{
\item{out.model}{outputs from model 1 (\code{MC}) or model 2 (\code{FWH})}

\item{analysis}{"experimental.design", convergence" or "posteriors". If NULL, the three are done.}

\item{nb_parameters_per_plot}{The number of parameters per plot to facilitate the visualisation}
}
\value{
The function returns a list with

\itemize{
\item "experimental_design" : a plot representing the presence/abscence matrix of GxE combinaisons.

\item "convergence" : a list with the plots of trace and density to check the convergence of the two MCMC only for chains that are not converging thanks to the Gelman-Rubin test. If all the chains converge, it is NULL

\item "posteriors" a list with
\itemize{

 \item for model 1
 \itemize{
   \item "sigma_distribution" : the distribution of the sigma is displayed on the Inverse Gamma distribution
   \item "parameter_posteriors" : a caterpillar plot is display for each mu_ij, beta_jk for each environment and for sigma_j
   \item "standardized_residuals" : a plot to check the normality of the residuals
 }

 \item for model 2
   \itemize{
   \item "parameter_posteriors" : a list with caterpillar plot for each alpha_i, beta_i and theta_j
   \item "standardized_residuals" : a plot to check the normality of the residuals. If the model went well it should be between -2 and 2.
   }

 }

\item "MCMC" : a data fame resulting from the concatenation of the two MCMC for each parameter. This object can be used for further analysis. There are as many columns than parameters and as many rows than iterations/10 (10 being the thin value by default in the models).
}
}
\description{
\code{analyse.outputs} displays plots to see if the model went well based ont the Gelman-Rubin test and plots of posteriors distributions. It is important to run this step before going ahead in the analysis otherwise you may make mistakes in the interpretation of the results.
}
\details{
For analyse = "convergence", the test used is the Gelman-Rubin test.
It may take some times to run.
More details with ?\code{gelman.diag} from the \code{coda} package.
Note that for \code{gelman.diag}, the argument \code{multivariate = FALSE} is used.
If you wish exhaustive information, looked at \code{ggmcmc::ggmcmc} with \code{ggmcmc(out.model$MCMC)}.
But be careful with the size of your MCMC output which are often too big to be performed in R.

More information in  the vignette. Type vignette("PPBstats").
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{MC}}, \code{\link{FWH}}, \code{\link{get.mean.comparisons}}
}

