% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_mixture.R
\name{analyse_mixtures}
\alias{analyse_mixtures}
\title{Analysis of mixtures performances from \code{\link{check_model.fit_model_bh_intra_location}} object and \code{\link{format_data_PPBstats.data_mixture.R}}}
\usage{
analyse_mixtures(
  res_model,
  data_mixture,
  plot.type = NULL,
  table.type = NULL,
  person = NULL,
  nb_parameters_per_plot = 15
)
}
\arguments{
\item{res_model}{output from \code{\link{check_model.fit_model_bh_intra_location}} function}

\item{data_mixture}{output from \code{\link{format_data_PPBstats.data_mixture.R}} function
A list of 3 elements
\itemize{
    \item Mixture_query
    \item Mixture_classic
    \item Mixture_S
}}

\item{plot.type}{the type of plot wanted. Default is NULL. Can be a vector of parameters
\itemize{
 \item "comp.in.farm" to compare mixtures to each of their components and the mean of their components in a farm
 \item "mix.comp.distribution" to get the histogram of the overyieldings, 
 the overyielding for each mixture being computed as the value of the mixture minus the mean value of its components, divided by the mean value of the mixture's component 
 \item "mix.gain.distribution" to get, for each mixture, the value of the lower component, the highest component and the mean of the components
 la suite --> expe mélange uniquement ===> à retirer
 \item "comp.mod" to compare the selection modalities of a mixture  
 \item "comp.mod.network" to get, for all mixtures, the comparison of the different selection modalities with the non selected mixture and the comparison of the different selection modalities
 \item "all" to get all the plots
}}

\item{table.type}{type of table to be returned. Default is NULL. Can be a vector of parameters
\itemize{
 \item \code{global.info} to get a table containing 
 \item \code{distribution} to get a table containing :
 \itemize{
    \item the number of mixtures tested
    \iem the proportion of mixtures which value exceeds the predicted value based on the mean of components' values
    \item the number of mixtures for which the mixture significantly exceeds the predicted value (alpha = 0.05)
    \item the proportion of mixtures which value is lower than its lowest component
    \item the proportion of mixtures which value is higher than its highest component
    \item mean value of all components
    \item mean value of all mixtures
    \item mean overyielding over all the mixtures tested. The overyielding is calculated as : 
    \deqn{\frac{1}{N} \sum\limits_{i=1}^N \frac{\overline{Y_{Mixture_{i}}} - \overline{Y_{Mean of components_{i}}}}{\overline{Y_{Mean of components_{i}}}} }
    \item standard deviation of overyieldings
    \item pvalue of the comparison test mean overyielding vs 0
 }
 \item \code{correlations} to get a table containing:
 \itemize{
   \item the correlation of the overyielding with: the number of components in the mixture and the variance observed among components
   \item the pvalue of the correlation
 }
 
 la suite --> expe mélange uniquement ===> à retirer
 \item \code{selection.modalities} (focusing on phenotypic mean) and \code{varIntra} (focusing on phenotypic variability) to get tables containing
 the comparison of the selected mixture and the non-selected mixture for all selection modalities, and the comparison between the different selection modalities:
   \itemize{
      \item mean value of the differences over all the mixtures tested
      \item standard deviation of differences
      \item statistic value for the comparison test (t.test if normal distribution of values, wilcox.test otherwise)
      \item pvalue of the comparison test
      \item significance symbol
      \item test used for the comparison
   }
 \item \code{all} to get all tables 
}}

\item{person}{if plot.type = "comp.in.farm" or "comp.mod", the farmers you want the analysis done for}

\item{nb_parameters_per_plot}{the number of parameters per plot}
}
\value{
the model results for the germplasm, environment and year wanted
}
\description{
Analysis of mixtures performances from \code{\link{check_model.fit_model_bh_intra_location}} object and \code{\link{format_data_PPBstats.data_mixture.R}}
}
\seealso{
\code{\link{PPBstats::analyse.outputs}}, \code{\link{shinemas2R::get.data}}
}
\author{
Gaelle Van Frank
}
