% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_data.check_model_GxE .R
\name{biplot_data.check_model_GxE}
\alias{biplot_data.check_model_GxE}
\title{Compute ecovalence and format PCA results for biplot}
\usage{
\method{biplot_data}{check_model_GxE}(x)
}
\arguments{
\item{x}{Output from \code{\link{check_model}} with \code{\link{model_GxE}}}
}
\value{
The function returns a list with:
\itemize{
\item info : a list with variable and gxe_analysis
\item data_ecovalence : the ecovalence matrix
\item data_interaction : the interaction matrix
\item pca : the pca object
}
}
\description{
\code{biplot_data.check_model_GxE} computes ecovalence and format PCA results from \code{\link{check_model}} with \code{\link{GxE}}
}
\details{
The ecovalence is the interaction matrix squared.
}
\seealso{
\itemize{
\item \code{\link{check_model.fit_model_GxE}}, 
\item \code{\link{plot.biplot_GxE}}
}
}
\author{
Pierre Riviere
}
