% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_data.check_model_hedonic.R
\name{biplot_data.check_model_hedonic}
\alias{biplot_data.check_model_hedonic}
\title{Format data for CA biplot}
\usage{
\method{biplot_data}{check_model_hedonic}(x)
}
\arguments{
\item{x}{Output from \code{\link{check_model}} with \code{\link{model_hedonic}}}
}
\value{
The function return a CA object
}
\description{
\code{biplot_data.check_model_hedonic} format data for CA plot from \code{\link{check_model}} with \code{\link{model_hedonic}}
}
\seealso{
\itemize{
\item \code{\link{check_model.fit_model_hedonic}}, 
\item \code{\link{plot.biplot_hedonic}}
}
}
\author{
Pierre Riviere
}
