% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_data.R
\name{biplot_data}
\alias{biplot_data}
\title{Format data for biplot representation}
\usage{
biplot_data(x)
}
\arguments{
\item{x}{outputs from 
\itemize{
 \item \code{\link{check_model.fit_model_GxE}}
 \item \code{\link{check_model.fit_model_hedonic}}
 \item \code{\link{check_model.fit_model_napping}}
 }}
}
\description{
\code{biplot_data} formats data for biplot representation
}
\details{
S3 method.
See for more details :
\itemize{
 \item \code{\link{biplot_data.check_model_GxE}} regarding \code{\link{model_GxE}}
 \item \code{\link{biplot_data.check_model_hedonic}} regarding \code{\link{model_hedonic}}
 \item \code{\link{biplot_data.check_model_napping}} regarding \code{\link{model_napping}}
}
}
