% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{check_analysis_argument}
\alias{check_analysis_argument}
\title{error message regadring check_model of bayesian analysis}
\usage{
check_analysis_argument(analysis)
}
\arguments{
\item{analysis}{type of analysis : "experimental_design", "convergence" or "posteriors"}
}
\description{
error message regadring check_model of bayesian analysis
}
