% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{check_convergence}
\alias{check_convergence}
\title{check convergence of bayesian model}
\usage{
check_convergence(out.model, model_name = "model1")
}
\arguments{
\item{out.model}{output from bayesian model}

\item{model_name}{name of the model}
}
\description{
check convergence of bayesian model
}
