% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{check_data_vec_variables}
\alias{check_data_vec_variables}
\title{check if variable are part of the data}
\usage{
check_data_vec_variables(data, vec_variables)
}
\arguments{
\item{data}{data frame}

\item{vec_variables}{variables to check}
}
\description{
check if variable are part of the data
}
