% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{check_list_out_check_model}
\alias{check_list_out_check_model}
\title{Check constitency of list with outputs from \code{\link{check_model}}}
\usage{
check_list_out_check_model(valid_models, list_out_check_model)
}
\arguments{
\item{valid_models}{A vector with valid models expected}

\item{list_out_check_model}{A list whose elements are output from \code{\link{check_model}}}
}
\value{
If the list do not contain valid models and all the elements of the list are coming from the same model, an error is returned.

If no error message are displayed the function returns a vector with the class of elements in list_out_check_model
}
\description{
\code{check_list_out_check_model} checks consistency of list with outputs from \code{\link{check_model}}
}
\author{
Facundo Munoz and Pierre Riviere
}
