% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.fit_model_bh_variance_intra.R
\name{check_model.fit_model_bh_variance_intra}
\alias{check_model.fit_model_bh_variance_intra}
\title{Check if the Hierarchical Bayesian variance intra model went well}
\usage{
\method{check_model}{fit_model_bh_variance_intra}(x)
}
\arguments{
\item{x}{outputs from \code{\link{model_bh_variance_intra}}}
}
\value{
It returns a list with the following elements:

\itemize{
 \item MCMC : a data fame resulting from the concatenation of the two MCMC for each parameter
 \item MCMC_conv_not_ok : a data fame resulting from the concatenation of the two MCMC for each parameter for environment where  some parameters did not converge for mu and beta
 \item data_ggplot a list containing information for ggplot:
 \itemize{
  \item mu
  \item sigma
  \item epsilon
 }
}
}
\description{
\code{check_model.fit_model_bh_variance_intra} computes tests to assess if the model went well. 
It is important to run this step before going ahead with the analysis otherwise you may make mistakes in the interpretation of the results.
}
\details{
S3 method.
The different test apply to the model are explained in the book \href{https://priviere.github.io/PPBstats_book/intro-agro.html#section-bayes}{here}.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{plot.check_model_bh_variance_intra}}
\item \code{\link{mean_comparisons}}
\item \code{\link{mean_comparisons.check_model_bh_variance_intra}}
}
}
\author{
Gaelle Van Frank and Pierre Riviere
}
