% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.fit_model_napping.R
\name{check_model.fit_model_napping}
\alias{check_model.fit_model_napping}
\title{Check if the napping model went well}
\usage{
\method{check_model}{fit_model_napping}(x)
}
\arguments{
\item{x}{outputs from \code{\link{model_napping}}}
}
\value{
It returns a list with the same elements as \code{\link{model_napping}} as the MFA object is already computed
}
\description{
\code{check_model.fit_model_napping} computes tests to assess if the model went well. 
It is important to run this step before going ahead with the analysis otherwise you may make mistakes in the interpretation of the results.
}
\details{
S3 method.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{plot.check_model_napping}}
\item \code{\link{biplot_data}}
\item \code{\link{biplot_data.check_model_napping}}
}
}
\author{
Pierre Riviere
}
