% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check if the model went well}
\usage{
check_model(out_model)
}
\arguments{
\item{out_model}{outputs from GxE, model_1, model_2}
}
\value{
The function returns a list depending on check_model input

\itemize{

\item GxE
\itemize{
 \item GxE the output from the model
 \item data_ggplot a list containing information for ggplot:
 \itemize{
  \item data_ggplot_residuals a list containing :
   \itemize{
    \item data_ggplot_normality
    \item data_ggplot_skewness_test
    \item data_ggplot_kurtosis_test
    \item data_ggplot_qqplot
    }
  \item data_ggplot_variability_repartition_pie
  \item data_ggplot_var_intra
 }
}


\item model_1
\itemize{
 \item MCMC
 \item MCMC_conv_not_ok
 \item data_env_with_no_controls
 \item data_env_whose_param_did_not_converge : a list with data frame with environments where some parameters did not converge for mu and beta.
 \item data_ggplot a list containing information for ggplot:
 \itemize{
  \item sigma_j
  \item mu_ij
  \item beta_jk
  \item sigma_j_2
  \item epsilon_ijk
 }
}


\item model_2
\itemize{
 \item MCMC
 \item MCMC_conv_not_ok
 \item model2.presence.abscence.matrix : a matrix germplasm x environment with the number of occurence in the data used for the model (i.e. with at least two germplasm by environments.)
 \item data_ggplot a list containing information for ggplot:
 \itemize{
  \item alpha
  \item beta
  \item theta
  \item epsilon
 }
}

}
}
\description{
\code{check_model} compute test to assess if model went well.
It is important to run this step before going ahead with the analysis otherwise you may make mistakes in the interpretation of the results.
}
\details{
The function calls one function: check_model_GxE, check_model_1, check_model_2

For model_1 and model_2, the test used for convergence is the Gelman-Rubin test.
It may take some time to run.
More details with ?\code{gelman.diag} from the \code{coda} package.
Note that for \code{gelman.diag}, the argument \code{multivariate = FALSE} is used.


MCMC outputs are a data fame resulting from the concatenation of the two MCMC for each parameter.
This object can be used for further analysis.
There are as many columns than parameters and as many rows than iterations/10 (10 being the thin value by default in the models).
The MCMC contains only parameters that converge thanks to the Gelman-Rubin test (if it has been done).
For model 1, all environments where at least one parameter do not converge are deleted.

The outputs of the function is used in
\itemize{
 \item mean_comparisons for GxE, model_1 and model_2
 \item parameter_groups for GxE and model_2
 \item predict_the_past_model_2 for model_2
}
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{GxE}}, \code{\link{model_1}}, \code{\link{model_2}}, \code{\link{check_model_GxE}}, \code{\link{check_model_1}}, \code{\link{check_model_2}}, \code{\link{mean_comparisons}}
}

