% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model_GxE.R
\name{check_model_GxE}
\alias{check_model_GxE}
\title{Check if the GxE model went well}
\usage{
check_model_GxE(out_GxE)
}
\arguments{
\item{out_GxE}{outputs from \code{\link{GxE}}}
}
\value{
See \code{\link{check_model}}
}
\description{
\code{check_model_GxE} computes tests to assess if the GxE model went well
}
\details{
See \code{\link{check_model}}
}
\seealso{
\itemize{
\item \code{\link{GxE}},
\item \code{\link{check_model}}
}
}

