% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model_model_2.R
\name{check_model_model_2}
\alias{check_model_model_2}
\title{Check if the model_2 model went well}
\usage{
check_model_model_2(out_model_2)
}
\arguments{
\item{out_model_2}{outputs from model_2 function}
}
\value{
See check_model
}
\description{
\code{check_model_model_2} compute test to assess if the model_2 model went well
}
\details{
See check_model
}
\seealso{
\code{\link{model_2}}, \code{\link{check_model}}
}

