% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_model_variance_intra.R
\name{check_model_model_variance_intra}
\alias{check_model_model_variance_intra}
\title{Check if the model_variance_intra model went well}
\usage{
check_model_model_variance_intra(out_model_varintra)
}
\arguments{
\item{out_model_variance_intra}{outputs from \code{\link{model_variance_intra}}}
}
\value{
See \code{\link{check_model}}
}
\description{
\code{check_model_variance_intra} computes tests to assess if the model_variance_intra model went well
}
\details{
See \code{\link{check_model}}
}
\seealso{
\itemize{
\item \code{\link{model_variance_intra}}, 
\item \code{\link{check_model}}
}
}
