% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{common_functions}
\alias{common_functions}
\alias{check_data_vec_variables}
\title{Some functions used in one or several functions of PPBstats}
\usage{
check_data_vec_variables(data, vec_variables)
}
\arguments{
\item{data}{data frame}

\item{vec_variables}{variables to check}
}
\description{
This file group functions used in several functions of PPBstats
}
\author{
Pierre Riviere

check if variable are part of the data
}
