% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comp.parameters.R
\name{comp.parameters}
\alias{comp.parameters}
\title{Get parameter comparisons two by two or to a given threshold based on MCMC outputs}
\usage{
comp.parameters(MCMC, parameter = "mu", type = 1, threshold = 1)
}
\arguments{
\item{MCMC}{MCMC data frame.}

\item{parameter}{The parameter on which to get the comparisons}

\item{type}{The type of comparisons. 1 for comparison two by two. 2 for comparison to a specific threshold.}

\item{threshold}{For type = 2, the threshold to which a parameter is different}
}
\value{
The function returns
\itemize{
\item for type = 1 : a square matrix with the probability of having a common distribution for each pair of parameter.
\item for type = 2 : a vector with the probability of being different from the given threshold.
}
}
\description{
\code{comp.parameters} performs parameter comparisons based on MCMC outputs
}
\details{
The comparisons is based on the probability of having a common distribution for each pair of parameters (type 1) or to be different from a specific threshold (type 2). More details in the vignette (type vignette ("PPBstats")).
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{check_model}}, \code{\link{mean_comparisons}}
}

