% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_agro_HA}
\alias{data_agro_HA}
\title{Data to study home and away}
\format{A data frame with 27 rows and 13 columns}
\usage{
data(data_agro_HA)
}
\description{
A dataset containing values for the following factors: seed_lot, location, year, germplasm, block, X, Y, group and version, Y1, Y1$date, y2, y2$date.
group represents home or away
version represents the location where come the germplasm from.
}
\author{
Pierre Rivière
}
\keyword{datasets}
