% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_agro_SR_4}
\alias{data_agro_SR_4}
\title{data set to study response to selection from data set data_model_bh_intra_location}
\format{A data frame with 998 rows and 12 columns}
\usage{
data(data_agro_SR_4)
}
\description{
data_model_bh_intra_location adapted with three more columns: expe_id, group and version
}
\details{
tkw is in grams
}
\author{
Pierre Rivière
}
\keyword{datasets}
