% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_hedonic}
\alias{data_hedonic}
\title{data coming from tomatoes hedonic test:}
\format{
A data frame with 244 rows and 11 columns
}
\usage{
data(data_hedonic)
}
\description{
A dataset containing 1 location and 6 germplasms.
The data set contain the following variables: sample, germplasm, location, juges, note, descriptors, Age, Sexe, 
Bio.Non.Bio, Circuit, Departement
}
\author{
Pierre Rivière
}
\keyword{datasets}
