% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_model_1}
\alias{data_model_1}
\title{Simulated data set for housand kernel weight (tkw) of bread wheat to test model_1 model}
\format{A data frame with 1574 rows and 11 columns}
\usage{
data(data_model_1)
}
\description{
A dataset containing one variable (tkw) and values for the following factors: location, year, germplasm, block, X and Y. 
The different values of the parameters to create the dataset are also in columns: mu_ij, beta_jk, epsilon_ijk and sigma_j
}
\details{
tkw is in grams
}
\author{
Pierre Rivière
}
\keyword{datasets}
