% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_model_variance_intra}
\alias{data_model_variance_intra}
\title{Simulated data set for spike weight of bread wheat to test model_variance_intra}
\format{A data frame with 1182 rows and 10 columns}
\usage{
data(data_model_variance_intra)
}
\description{
A dataset containing one variable (spike weight) 
and values for the following factors: location, year, germplasm, block, X and Y. 
The different values of the parameters to create the dataset are also in columns: mu_ijk, sigma_ij and epsilon_ijkl
}
\details{
spike weight is in grams
}
\author{
Gaelle Van Frank
}
\keyword{datasets}
