% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_napping}
\alias{data_napping}
\title{data coming from tomatoes hedonic test:}
\format{A data frame with 63 rows and 6 columns}
\usage{
data(data_napping)
}
\description{
A dataset containing 1 location and 9 germplasms.
The data set contain the following variables: juges, X, Y, descriptors, germplasm, location
}
\author{
Pierre Rivière
}
\keyword{datasets}
