% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_network_bipart}
\alias{data_network_bipart}
\title{Simulated data set for bipart network which represents "which location has which germplasm which year":}
\format{A data frame with 73 rows and 14 columns}
\usage{
data(data_network_bipart)
}
\description{
A dataset containing 3 locations, 11 germplasms and 5 years.
The data set contain the following variables: germplasm, location, year, alt, long, lat

The data set have been generated from data_network_unipart_sl with relation diffusion or reproduction

See the vignette for more information.
}
\author{
Pierre Rivière
}
\keyword{datasets}
