% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_network_unipart_location}
\alias{data_network_unipart_location}
\title{Simulated data set for unipart network which represents relation of germplasm diffusion between location}
\format{
A data frame with 22 rows and 14 columns
}
\usage{
data(data_network_unipart_location)
}
\description{
A dataset containing four locations, 15 germplasms and 5 years.
Relation between seed lots concern diffusion.
The data set contain the following variables: location_parent, location_child, 
relation_year_start, relation_year_end, 
germplasm_parent, year_parent, alt_parent, long_parent, lat_parent, 
germplasm_child, year_child, alt_child, long_child, lat_child

The data set have been generated from data_network_unipart_sl

See the vignette for more information.
}
\author{
Pierre Rivière
}
\keyword{datasets}
