% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_network_unipart_sl}
\alias{data_network_unipart_sl}
\title{Simulated data set for unipart network for seed lots which represents relation between between seed lots}
\format{
A data frame with 104 rows and 17 columns
}
\usage{
data(data_network_unipart_sl)
}
\description{
A dataset containing four locations, 15 germplasms and 5 years.
Relation between seed lots are diffusion, mixture, reproduction and selection.
The data set contain the following variables: seed_lot_parent, seed_lot_child, 
relation_type, relation_year_start, relation_year_end, 
germplasm_parent, location_parent, year_parent, alt_parent, long_parent, lat_parent, 
germplasm_child, location_child, year_child, alt_child, long_child, lat_child

See the vignette for more information.
}
\author{
Pierre Rivière
}
\keyword{datasets}
