% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_version_MR}
\alias{data_version_MR}
\title{Data to study migrant and residant}
\format{A data frame with 25 rows and 5 columns}
\usage{
data(data_version_MR)
}
\description{
A dataset containing values for the following factors: year, location, germplasm, group and version.
group represents migrant or residant
version represents the location where come the germplasm from.
}
\author{
Pierre Rivière
}
\keyword{datasets}
