% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_version_SR}
\alias{data_version_SR}
\title{Data to study response to selection based on data_version format}
\format{A data frame with 20 rows and 5 columns}
\usage{
data(data_version_SR)
}
\description{
A dataset containing values for the following factors: year, location, germplasm, group and version.
group represents differential selection (S) or reponse to selection (R)
version represents bouquet or vrac.
}
\author{
Pierre Rivière
}
\keyword{datasets}
