% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe the data set in order to choose the appropriate analysis to carry out}
\usage{
describe_data(data, vec_variables, nb_parameter_per_grid = 5)
}
\arguments{
\item{data}{}

\item{vec_variables}{Vector of variables to describe}

\item{nb_parameter_per_grid}{Nomber of parameter on each histogram on the gird}
}
\value{
The function returns a list with for each environment, the estimated value of the germplasms that were not grown in this environment.

The function returns a list with, for each variable, a list with :
\itemize{
 \item A presence.abscence plot
 \item Histogram of the variable
 \item Histogram of the variable per germplasm
 \item Histogram of the variable per location
 \item Histogram of the variable per year
}
}
\description{
\code{describe_data} describes the data set in order to choose the appropriate analysis to carry out
}
\author{
Pierre Riviere
}

