% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_experiment.R
\name{design_experiment}
\alias{design_experiment}
\title{Provides experimental design for the different situations corresponding to 
the choosen family of analysis}
\usage{
design_experiment(
  expe.type,
  location = "location",
  year = "2016",
  germplasm,
  controls,
  nb.controls.per.block,
  nb.blocks,
  nb.cols,
  return.format = "standard",
  allow_XXXX = TRUE
)
}
\arguments{
\item{expe.type}{The type of experiment to settle "satellite-farm", "regional-farm", "row-column", "fully-repicated", "IBD". See details for more information.}

\item{location}{Name of the location.}

\item{year}{Year of the experiment.}

\item{germplasm}{Vector with the names of the germplasm.}

\item{controls}{Vector with names of controls.}

\item{nb.controls.per.block}{Number of controls per blocks.}

\item{nb.blocks}{Number of blocks. For expe.type = "IBD", it corresponds to the number of locations.}

\item{nb.cols}{Number of columns in the design. The number of rows is computed automaticaly.}

\item{return.format}{"standard" or "shinemas". See details for more information.}

\item{allow_XXXX}{TRUE or FALSE. Allow to have XXXX placed in stead of germplasm in order to fill the design.}
}
\value{
The function returns a list with
\itemize{
 \item data.frame : A data frame according to return.format
 \item design : An image of the experimental design
 }
}
\description{
\code{design_experiment} provides experimental design for the different situations corresponding to 
the choosen family of analysis
}
\details{
regarding return.format :
\itemize{
\item  "standard" return a data frame with the following columns : "location", "year", "germplasm", "block", "X" and "Y".
\item "shinemas" for SHiNeMaS reproduction template file. It returns a data frame with the following columns :"project", "sown_year", "harvested_year", "id_seed_lot_sown", "intra_selection_name", "etiquette", "split", "quantity_sown", "quantity_harvested", "block", "X" and "Y".
}

Note that an efficient R package to design experiment is DiGGer, see \url{http://www.austatgen.org/software/} for mor details. Unfortunately, the lience of the package did not allow us to fork the code. 

In this function, the code is based on the following algorithms:
\itemize{
 \item "satellite-farm"
 \enumerate{
 \item randomize the germplasm
  \item get the data frame with one block, two columns and one control in each column
 \item randomize the columns
 }
 \item "regional-farm"
 \enumerate{
 \item randomize the germplasm
  \item get the data frame with blocks, columns and controls in each block
  \item arrange at least one control per row
  \item for each row, put control in different column
  \item randomize rows and columns
  \item check that controls do not touch each other
  \item check number of controls in col and row and send a warning message if control are missing in rows or columns
 }
 \item "row-column"
 Note that expe.type = "row-column" is particular case of expe.type = "regional-farm" where the number of controls must be at least the number of columns or rows.
 \enumerate{
 \item randomize the germplasm
  \item get the data frame with blocks, columns and controls in each block
  \item arrange at least one control per row
  \item for each row, put control in different column
  \item randomize rows and columns
  \item check that controls do not touch each other
  \item check number of controls in col and row and send a stop message if control are missing in rows or columns
 }
 \item "fully-repicated"
 \enumerate{
 \item randomize the germplasm
  \item get the data frame with blocks, columns and germplasm in each block
  \item arrange randomization so that no germplasm can be in the same column between blocks
 }
 \item "IBD"
the randomization is based on the ibd function in the ibd package. See \code{?ibd} for more information
}

See exampels in the book : \href{https://priviere.github.io/PPBstats_book/doe.html}{here}.
}
\author{
Pierre Riviere
}
