% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{extra_functions}
\alias{check_data_vec_variables}
\alias{extra_functions}
\title{Some functions used in several functions of PPBstats}
\usage{
check_data_vec_variables(data, vec_variables)
}
\description{
This file group several functions used in several functions of PPBstats:
\itemize{
	\item describe_data
}
}
\author{
Pierre Riviere
}

