% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.data_agro_LF.R
\name{format_data_PPBstats.data_agro_LF}
\alias{format_data_PPBstats.data_agro_LF}
\title{Check and format the data to be used by PPBstats functions for data agro LF to study local adaptation.}
\usage{
format_data_PPBstats.data_agro_LF(data)
}
\arguments{
\item{data}{The data frame to format.
It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "group", "version", "..."), 
with "..." the variables.
The variables can be linked to their corresponding dates. 
The dates are associated to their corresponding variable by $.
For example the date associated to variable y1 is y1$date.
The date must have format year-month-day, e.g. 2017-12-05}
}
\description{
\code{format_data_PPBstats.data_agro_LF} checks and formats the data to be used by PPBstats functions for agro HA to study local adpatation
}
\details{
\itemize{
 \item group: the location where the germplasm come from
 \item version: foreign or local
}

More details can be found in the book \href{https://priviere.github.io/PPBstats_book/local-foreign.html}{here}.
}
\seealso{
\code{\link{format_data_PPBstats}}
}
\author{
Pierre Riviere
}
