% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.data_agro_SR.R
\name{format_data_PPBstats.data_agro_SR}
\alias{format_data_PPBstats.data_agro_SR}
\title{Check and format the data to be used by PPBstats functions for data agro SR to study response to selection.}
\usage{
format_data_PPBstats.data_agro_SR(data)
}
\arguments{
\item{data}{The data frame to format.
It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "expe_id", "group", "version", "..."), 
with "..." the variables.
The variables can be linked to their corresponding dates. 
The dates are associated to their corresponding variable by $.
For example the date associated to variable y1 is y1$date.
The date must have format year-month-day, e.g. 2017-12-05}
}
\description{
\code{format_data_PPBstats.data_agro_SR} checks and formats the data to be used by PPBstats functions for agro SR to study response to selection
}
\details{
\itemize{
 \item expe_id: an id with
  \itemize{
   \item S couple and R couple (i.e. four rows) OR
   \item S couple (i.e. two rows) OR
   \item R couple (i.e. two rows)
  }
 \item group: S or R
 \item version: bouquet or vrac
}

expe_id is useful for example if there are several selection in one germplasm or if there are several origin for a given germplasm

More details explaining these two particular cases can be found in the book \href{https://priviere.github.io/PPBstats_book/response-to-selection.html}{here}.
}
\seealso{
\code{\link{format_data_PPBstats}}
}
\author{
Pierre Riviere
}
