% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.data_mixture.R
\name{format_data_PPBstats.data_mixture}
\alias{format_data_PPBstats.data_mixture}
\title{Check and format the data to be used by PPBstats functions for mixture analyses}
\usage{
format_data_PPBstats.data_mixture()
}
\arguments{
\item{data}{The data frame to format, see details.}
}
\value{

}
\description{
\code{format_data_PPBstats} checks and formats the data to be used by PPBstats functions for mixture analyses
}
\seealso{
\code{\link{format_data_PPBstats}}
}
\author{
Gaelle van Frank
}
