% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.data_organo_napping.R
\name{format_data_PPBstats.data_organo_napping}
\alias{format_data_PPBstats.data_organo_napping}
\title{Check and format the data to be used by PPBstats functions for napping analyses}
\usage{
format_data_PPBstats.data_organo_napping(data, threshold)
}
\arguments{
\item{data}{The data frame with the following columns: sample, juges, X, Y, descriptors, germplasm, location.
The descriptors must be separated by ";"}

\item{threshold}{number of occurence of descriptors <= threshold are kept}
}
\description{
\code{format_data_PPBstats.data_organo_napping} checks and formats the data to be used by PPBstats functions for napping analyses
}
\details{
See the book for more details \href{https://priviere.github.io/PPBstats_book/napping.html#format-the-data-9}{here}.
}
\seealso{
\code{\link{format_data_PPBstats}}
}
\author{
Pierre Riviere
}
