% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.R
\name{format_data_PPBstats}
\alias{format_data_PPBstats}
\title{Check and format the data to be used in PPBstats functions}
\usage{
format_data_PPBstats(data, type = "data_agro", code, threshold = NULL,
  network_part = NULL)
}
\arguments{
\item{data}{The data frame. See details.}

\item{type}{type of format : 
\itemize{
 \item data_network
 \item data_agro
 \item data_agro_version
 \item data_organo_napping
 \item data_organo_hedonic
 }
See details.}

\item{code}{For type = data_organo, data frame with code of samles. See details.}

\item{threshold}{For type = data_organo, number of occurence of descriptors <= threshold are kept}

\item{network_part}{For type = "data_network", element of the network, it can
\itemize{
 \item for unipartite network : "seed_lots"
 \item for bipartite network : 
 \itemize{
  \item c("germplasm", "location") or
  \item c("location", "germplasm") or
  \item c("germplasm", "year") or
  \item c("year", "germplasm") or
  \item c("location", "year") or
  \item c("year", "location")
 }
}}
}
\value{
The function returns the data with the right format (i.e. class)

For "data_network", it return a igraph object coming from igraph::graph_from_data_frame()
}
\description{
\code{format_data_PPBstats} checks and formats the data to be used in PPBstats functions
}
\details{
\itemize{
 \item For type = "data_network" : The data frame are different regarding type of network
 \itemize{
  \item for unipartite network, the data must have the following columns : 
  "seed_lot_parent", "seed_lot_child", "relation_type",
  "germplasm_parent", "location_parent", "year_parent",
  "germplasm_child", "location_child", "year_child".
  
  It can have in option : "alt_parent", "long_parent", "lat_parent", 
  "alt_child", "long_child", "lat_child" to get map representation
  
  It can have supplementary variables with tags "_parent", "_child" or "_relation".
  
  
  \item for bipartite network, the data must have the following columns : "germplasm", "location", "year"
  
  It can have in option : "alt", "long", "lat" to get map representation
 }
 
 
 \item For type = "data_agro" : 
 It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.
 The variables can be linked to their corresponding dates. 
 The dates are associated to their corresponding variable by $.
 For example the date associated to variable y1 is y1$date.
 The date must have format year-month-day, e.g. 2017-12-05
 
 \item For type = "data_agro_version" : 
 It should have the following columns: c("year", "germplasm", "location", "group", "version").
 The group refers to an id that contains two different versions.
 For example for group 1, there is version 1 and 2. 
 See data(data_version) for an example.
 
 \item For type = "data_organo_napping"
 \itemize{
  \item data is a data frame with the following columns: sample, juges, X, Y, descriptors. 
  The descriptors must be separated by ";"
  \item code is data frame with the following columns germplasm, location, code.
  The function merge data and code to join the information.
 }

 \item For type = "data_organo_hedonic"
 
}
}
\author{
Pierre Riviere
}
