% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.at.least.X.groups.R
\name{get.at.least.X.groups}
\alias{get.at.least.X.groups}
\title{Get the value of type one error needed to have X groups.}
\usage{
get.at.least.X.groups(Mpvalue, MCMC, p.adj = "soft.bonf", precision = 5e-04)
}
\arguments{
\item{Mpvalue}{A vector or a matrix with the probabilities for two parameters to be equal. It comes from the function \code{\link{comp.parameters}}}

\item{MCMC}{MCMC. It is a data frame.}

\item{p.adj}{NULL for no adjustement of the type one error. "soft.bonf" for a soft bonferonni correction to take into account multiple comparisons (alpha / number of parameters).}

\item{precision}{The precision of the type one error with the corresponding groups. The smaller the better, but the smaller the more time consuming due to computing matters}
}
\value{
The function returns a vector with the value of type one error for each group
}
\description{
\code{get.at.least.X.groups} gets the value of type one error needed to have X groups. X goes from 2 to the maximum number of groups possible.
}
\details{
This function is helpful to get at least two (or more) groups with a given type one error.

When there is no data, it returns NA. It may depend on the precision. For exemple, if you go from 5 groups to 2 groups, it will be NA for 4 groups and 3 groups because the precision was not small enought.
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
\item \code{\link{mean_comparisons_model_1}}
\item \code{\link{mean_comparisons_model_2}}
}
}

