% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get.caterpillar.plot}
\alias{get.caterpillar.plot}
\title{get caterpillar plot to view posterior of bayesian model}
\usage{
get.caterpillar.plot(x, xmin, xmax)
}
\arguments{
\item{x}{data frame}

\item{xmin}{xmin of the plot}

\item{xmax}{xmax of the plot}
}
\description{
get caterpillar plot to view posterior of bayesian model
}
