% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.env.info.R
\name{get.env.info}
\alias{get.env.info}
\title{Get regional farms data and satellite farms data}
\usage{
get.env.info(D, nb_ind)
}
\arguments{
\item{D}{The data frame with its specific column names used in \code{\link{model_bh_intra_location}}}

\item{nb_ind}{The minimal number of individuals per variable}
}
\value{
The function returns a list with 
\itemize{
\item "vec_env_with_no_data": a vector with the environments without data for the given variable
\item "vec_env_with_no_controls": a vector with the environments with no controls
\item "vec_env_with_controls": a vector with the environments with controls
\item "vec_env_RF": a vector with the environments as regional farms (i.e. with at least two blocks)
\item "vec_env_SF": a vector with the environments as satellite farms (i.e. with one block)
\item "D_RF": a data frame with regional farms
\item "D_SF": a data frame with satellite farms
}
}
\description{
\code{get.env.info} look at a data set and separate regional farms and satellite farms in two datasets. This function is used in \code{\link{model_bh_intra_location}}.
}
\seealso{
\code{\link{model_bh_intra_location}}
}
\author{
Pierre Riviere
}
