% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot.R
\name{get.ggplot}
\alias{get.ggplot}
\title{Get ggplot objects to visualize outputs from the analysis}
\usage{
get.ggplot(data, ggplot.type = "interaction", nb_parameters_per_plot = 8)
}
\arguments{
\item{data}{The data to plot. It can come from \code{get.mean.comparison}, \code{get.parameter.groups}, \code{predict.the.past}, \code{MC}$data_env_with_no_controls or \code{analye.outputs$model1.data_env_whose_param_did_not_converge}.}

\item{ggplot.type}{The type of plot you wish:
\itemize{
 \item from \code{get.mean.comparison}
 \itemize{
   \item from model 1 (\code{MC}) :  "barplot", "interaction", "score"
   \item from model 2 (\code{FWH}): "barplot"
   }

 \item from \code{get.parameter.groups} only from model 2 (\code{FWH}): "PCA"

 \item from \code{predict.the.past}$predict.the.past : "barplot", "interaction"

 \item from \code{MC}$data_env_with_no_controls : "barplot", "interaction"

 \item from \code{analye.outputs$model1.data_env_whose_param_did_not_converge} : "barplot", "interaction"

 }}

\item{nb_parameters_per_plot}{The number of parameters per plot to facilitate the visualization}
}
\value{
The function returns a list of ggplot objects
}
\description{
\code{get.ggplot} returns ggplot objects to visualize output from the analysis
}
\details{
\itemize{
\item From \code{get.mean.comparison} and from model 1 (\code{MC}) there are one plot per environment.
For ggplot.type = "interaction" and ggplot.type = "score", nb_parameters_per_plot is the number of entries (mu_ij).
This is meaningful if you have data for several years.

ggplot.type = "score" display a plot with a score according to which group the entry was allocated.
An high score means that the entry was in a group with an high mean.
A low score means that the entry was in a group with an low mean.
\item From \code{get.parameter.groups} from model 2 (\code{FWH}) "PCA" display the PCA and the groups of parameters.
}

On each plot, the alpha value and the alpha correction are displayed.
alpha = Imp means that no differences were possible to find.
For ggplot.type = "interaction" and ggplot.type = "score", it is display under the form: alpha | alpha correction
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{get.mean.comparisons}}, \code{\link{get.parameter.groups}}
}

