% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.PPBstats.data.R
\name{get.PPBstats.data}
\alias{get.PPBstats.data}
\title{Get PPBstats datas to run example of the vignette}
\usage{
get.PPBstats.data()
}
\arguments{
\item{folder_name}{name of the folder where the RData will be. "data" by default.}
}
\description{
\code{get.PPBstats.data} download Rdata from internet. This is because Rdata file are too big to be stored in the package.
}
\details{
It creates a PPBstats_data folder in the current directory and download the Rdata files
}
\author{
Pierre Riviere
}

