% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.significant.groups.R
\name{get.significant.groups}
\alias{get.significant.groups}
\title{Get significant groups of differences for a set of parameters based on MCMC outputs}
\usage{
get.significant.groups(Mpvalue, MCMC, alpha = 0.05, p.adj = NULL)
}
\arguments{
\item{Mpvalue}{Square matrix with the probability to have a common distribution for each pair of parameter. It comes from \code{comp.parameters}}

\item{MCMC}{MCMC outputs from \code{analysis.outputs}. It is a data frame.}

\item{alpha}{The level of type one error. 0.05 (5\%) by default}

\item{p.adj}{NULL For no adjustement of the pvalue. "soft.bonf" for a soft bonferonni correction to take into account multiple comparisons (alpha / nb of parameters).}
}
\value{
The function returns a data frame of four columns and (nb of parameters) rows. Columns are parameters, mean, median and groups
}
\description{
\code{get.significant.groups} gets significant groups of differences for a set of parameters based on MCMC outputs
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{analyse.outputs}}, \code{\link{get.mean.comparisons}}, \code{\link{get.ggplot}}
}

