% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get_biplot}
\alias{get_biplot}
\title{get ggplot object for a biplot based on output from FactoMineR::PCA}
\usage{
get_biplot(res.pca)
}
\arguments{
\item{res.pca}{output from FactoMineR::PCA}
}
\description{
get ggplot object for a biplot based on output from FactoMineR::PCA
}
