% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ggplot.R
\name{get_ggplot}
\alias{get_ggplot}
\title{Get ggplot objects to visualize output from several functions}
\usage{
get_ggplot(data, data_2 = NULL, data_version = NULL,
  ggplot.type = "interaction", nb_parameters_per_plot = 8)
}
\arguments{
\item{data}{Outputs from
\itemize{
 \item \code{check_model} or
 \item \code{mean_comparisons} or
 \item \code{parameter_groups} or
 \item \code{biplot_GxE} or
 \item \code{cross_validation_model_2} or
 \item \code{predict_the_past_model_2}
}}

\item{data_2}{Outputs from \code{mean_comparisons} from \code{check_model} from \code{model_2}.
This argument is used only when argument data comes from \code{mean_comparisons} from \code{check_model} from \code{model_2}.}

\item{data_version}{data set with the following columns: "year", "germplasm", "location", "group", "version". The group refers to an id that contains two different versions. For example for group 1, there is version 1 and 2. See data(data_version) for an example.
This argument is used only for \code{mean_comparisons} from
\itemize{
 \item \code{check_model} regarding \code{model_1}
 \item \code{check_model} regarding \code{predict_the_past_model_2}
}}

\item{ggplot.type}{The type of plot you wish.
\itemize{
 \item "barplot" for mean_comparisons from
 \itemize{
  \item check_model from GxE, model_1, model_2
  \item predict_the_past_model_2
  }
 \item "interaction.plot" for mean_comparisons from
 \itemize{
  \item check_model from model_1
  \item predict_the_past_model_2
  }
 \item "score" for mean_comparisons from
 \itemize{
  \item check_model from model_1
  \item predict_the_past_model_2
  }
 \item "biplot-alpha-beta" for mean_comparisons from check_model from model_2
 }

There are automatic plots generated for several functions:
\itemize{
 \item \code{biplot_GxE}
 \item \code{check_model}
 \item \code{parameter_groups}
 \item \code{cross_validation_model_2}
 }
See details for more information.}

\item{nb_parameters_per_plot}{The number of parameters per plot to facilitate the visualization}
}
\value{
to do
}
\description{
\code{get_ggplot} returns ggplot objects to visualize outputs from several functions
}
\details{
to do
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
 \item \code{check_model}
 \item \code{mean_comparisons}
 \item \code{parameter_groups}
 \item \code{biplot_GxE}
 \item \code{cross_validation_model_2}
 \item \code{predict_the_past_model_2}
}
}

