% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get_mcmc_traceplot_density}
\alias{get_mcmc_traceplot_density}
\title{get mcmc traceplot density to view posterior of bayesian model}
\usage{
get_mcmc_traceplot_density(MCMC)
}
\arguments{
\item{MCMC}{MCMC chain from bayesian model}
}
\description{
get mcmc traceplot density to view posterior of bayesian model
}
