% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get_mean_comparisons_and_Mpvalue}
\alias{get_mean_comparisons_and_Mpvalue}
\title{get mean comparisons and square matrix with pvalue from MCMC for bayesian models}
\usage{
get_mean_comparisons_and_Mpvalue(MCMC, parameter, type, threshold, alpha,
  p.adj, precision, get.at.least.X.groups)
}
\arguments{
\item{MCMC}{MCMC}

\item{parameter}{parameter}

\item{type}{type}

\item{threshold}{threshold}

\item{alpha}{alpha}

\item{p.adj}{p.adj}

\item{precision}{precision}

\item{get.at.least.X.groups}{get.at.least.X.groups}
}
\description{
get mean comparisons and square matrix with pvalue from MCMC for bayesian models
}
\details{
see \code{\link{mean_comparisons}}
}
