% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get_perpendicular_segment}
\alias{get_perpendicular_segment}
\title{get coordinate to draw perpendicular segment}
\usage{
get_perpendicular_segment(x1, y1, x2, y2, x3, y3, longer = FALSE)
}
\arguments{
\item{x1}{x coordinate of point 1}

\item{y1}{y coordinate of point 1}

\item{x2}{x coordinate of point 2}

\item{y2}{y coordinate of point 2}

\item{x3}{x coordinate of point 3}

\item{y3}{y coordinate of point 3}

\item{longer}{TRUE or FALSE}
}
\description{
get coordinate to draw perpendicular segment
}
\details{
following formulas thanks to jdbertron cf http://stackoverflow.com/questions/10301001/perpendicular-on-a-line-segment-from-a-given-point
}
