% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_models.R
\name{get_result_model}
\alias{get_result_model}
\title{Select parts of models results from PPBstats::analyse.outputs (eigher parameters comparisons or MCMC).}
\usage{
get_result_model(
  res_model,
  data,
  location = NULL,
  type_result = "comparison",
  variable,
  model,
  param = NULL,
  year = NULL
)
}
\arguments{
\item{res_model}{output from \code{model_1}, \code{model_2} or \code{model_varintra} function}

\item{data}{output from the \code{shinemas2R::get.data} function}

\item{location}{the location wanted}

\item{type_result}{the results wanted : parameters comparison ("comparison") or MCMC ("MCMC")}

\item{variable}{the variable}

\item{model}{the model from which the results were obtained : model_1, model_2, model_variance_intra}

\item{param}{The parameter wanted ("mu","beta","sigma" if model_1 is used, "alpha","beta","theta" if model_2 is used, "sigma" if model_variance_intra model is used)}

\item{year}{the year wanted}
}
\value{
the model results for the germplasm, environment and year wanted
}
\description{
Select parts of models results from PPBstats::analyse.outputs (eigher parameters comparisons or MCMC).
}
\seealso{
\code{\link{PPBstats::analyse.outputs}}, \code{\link{shinemas2R::get.data}}
}
\author{
Gaelle Van Frank
}
