% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_biplot_GxE.R
\name{ggplot_biplot_GxE}
\alias{ggplot_biplot_GxE}
\title{Get ggplot objects from biplot_GxE}
\usage{
ggplot_biplot_GxE(out_biplot_GxE)
}
\arguments{
\item{out_biplot_GxE}{outputs from biplot_GxE function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_biplot_GxE} returns ggplot objects from biplot_GxE
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{biplot_GxE}}
}

