% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_check_model_GxE.R
\name{ggplot_check_model_GxE}
\alias{ggplot_check_model_GxE}
\title{Get ggplot objects from check_model_GxE}
\usage{
ggplot_check_model_GxE(out_check_model, nb_parameters_per_plot = 10)
}
\arguments{
\item{out_check_model_GxE}{outputs from \code{\link{check_model_GxE}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_check_model_GxE} returns ggplot objects from \code{\link{check_model_GxE}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{check_model_GxE}}
}
}

