% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_check_model_model_1.R
\name{ggplot_check_model_model_1}
\alias{ggplot_check_model_model_1}
\title{Get ggplot objects from check_model_model_1}
\usage{
ggplot_check_model_model_1(out_check_model_1, nb_parameters_per_plot = 10)
}
\arguments{
\item{out_check_model_model_1}{outputs from check_model_model_1 function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_check_model_model_1} returns ggplot objects from check_model_model_1
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{check_model_model_1}}
}

