% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_check_model_model_variance_intra.R
\name{ggplot_check_model_model_variance_intra}
\alias{ggplot_check_model_model_variance_intra}
\title{Get ggplot from check_model_model_variance_intra}
\usage{
ggplot_check_model_model_variance_intra(out_check_model_varintra,
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_check_model_model_variance_intra}{outputs from \code{\link{check_model_model_variance_intra}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_check_model_model_variance_intra} returns ggplot from \code{\link{check_model_model_variance_intra}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{check_model_model_variance_intra}}
}
}

