% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_cross_validation_model_2.R
\name{ggplot_cross_validation_model_2}
\alias{ggplot_cross_validation_model_2}
\title{Get ggplot objects from cross_validation_model_2}
\usage{
ggplot_cross_validation_model_2(out_cross_validation_model_2)
}
\arguments{
\item{out_cross_validation_model_2}{outputs from cross_validation_model_2 function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_cross_validation_model_2} returns ggplot objects from cross_validation_model_2
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{cross_validation_model_2}}
}

