% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_discrimitiveness_vs_representativeness.R
\name{ggplot_discrimitiveness_vs_representativeness}
\alias{ggplot_discrimitiveness_vs_representativeness}
\title{Get "discrimitiveness vs representativeness" ggplot objects from PCA on interaction matrix}
\usage{
ggplot_discrimitiveness_vs_representativeness(res.pca)
}
\arguments{
\item{res.pca}{PCA object on interaction matrix}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_discrimitiveness_vs_representativeness} returns ggplot objects from PCA on interaction matrix
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{biplot_GxE}}
}

