% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_GxE.R
\name{ggplot_mean_comparisons_GxE}
\alias{ggplot_mean_comparisons_GxE}
\title{Get ggplot from mean_comparisons_GxE}
\usage{
ggplot_mean_comparisons_GxE(mean_comparisons_GxE, ggplot.type = "barplot",
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_mean_comparisons_GxE}{outputs from \code{\link{mean_comparisons_GxE}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_mean_comparisons_GxE} returns ggplot from \code{\link{mean_comparisons_GxE}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{mean_comparisons_GxE}}
}
}

