% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_model_1.R
\name{ggplot_mean_comparisons_model_1}
\alias{ggplot_mean_comparisons_model_1}
\title{Get ggplot objects from mean_comparisons_model_1}
\usage{
ggplot_mean_comparisons_model_1(out_mean_comparisons_model_1,
  data_version = NULL, ggplot.type = "interaction",
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_mean_comparisons_model_1}{outputs from mean_comparisons_model_1 function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_mean_comparisons_model_1} returns ggplot objects from mean_comparisons_model_1
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{mean_comparisons_model_1}}
}

