% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_predict_the_past_model_2.R
\name{ggplot_mean_comparisons_predict_the_past_model_2}
\alias{ggplot_mean_comparisons_predict_the_past_model_2}
\title{Get ggplot from mean_comparisons_predict_the_past_model_2}
\usage{

  ggplot_mean_comparisons_predict_the_past_model_2(out_mean_comparisons_predict_the_past_model_2,
  data_version = NULL, ggplot.type = "interaction",
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_mean_comparisons_predict_the_past_model_2}{outputs from \code{\link{mean_comparisons_predict_the_past_model_2}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_mean_comparisons_predict_the_past_model_2} returns ggplot from \code{\link{mean_comparisons_predict_the_past_model_2}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item\code{\link{mean_comparisons_predict_the_past_model_2}}
}
}

