% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_which_won_where.R
\name{ggplot_which_won_where}
\alias{ggplot_which_won_where}
\title{Get "which won where" ggplot objects from PCA on interaction matrix}
\usage{
ggplot_which_won_where(res.pca)
}
\arguments{
\item{res.pca}{PCA object on interaction matrix}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_which_won_where} returns ggplot objects from PCA on interaction matrix
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{biplot_GxE}}
}

