% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GxE_build_interaction_matrix.R
\name{GxE_build_interaction_matrix}
\alias{GxE_build_interaction_matrix}
\title{Compute interaction matrix}
\usage{
GxE_build_interaction_matrix(data, gxe_analysis)
}
\arguments{
\item{data}{data with at least the following columns: year, location, germplasm, block, X, Y}

\item{gxe_analysis}{the analysis to carry out: "AMMI" or "GGE"}
}
\value{
The function returns the interaction matrix
}
\description{
\code{GxE_build_interaction_matrix} computes interaction matrix for AMMI or GGE model
}
\details{
The computation is insoired from the agricolae::AMMI function. Another way to do has been explore but did not succeed regarding missing data. The code is inside the function.
}
\author{
Pierre Riviere
}
\seealso{
\code{GxE}
}

