% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_to_LF.R
\name{HA_to_LF}
\alias{HA_to_LF}
\title{Transform home away data to local foreign data}
\usage{
HA_to_LF(data_agro_HA)
}
\arguments{
\item{data_agro_HA}{data home away}
}
\value{
a data frame of class data_agro_LF
}
\description{
Transform home away data to local foreign data
}
\details{
change home to local and away to foreign
}
