% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{is.inside.sector}
\alias{is.inside.sector}
\title{to know if a point is inside an area}
\usage{
is.inside.sector(x, y, x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x}{x coordinate of point to know if it is inside the area or not}

\item{y}{y coordinate of point to know if it is inside the area or not}

\item{x1}{x coordinate of point 1 of the area}

\item{y1}{y coordinate of point 1 of the area}

\item{x2}{x coordinate of point 2 of the area}

\item{y2}{y coordinate of point 2 of the area}

\item{x3}{x coordinate of point 3 of the area}

\item{y3}{y coordinate of point 3 of the area}
}
\description{
to know if a point is inside an area
}
\details{
it is used for ggplot_which_won_where.R, ggplot_mean_vs_stability.R
}
