% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_to_HA.R
\name{LF_to_HA}
\alias{LF_to_HA}
\title{transform local foreign data to home away data}
\usage{
LF_to_HA(data_agro_LF)
}
\arguments{
\item{data_agro_LF}{data local foreign}
}
\value{
a data frame of class data_agro_HA
}
\description{
transform local foreign data to home away data
}
\details{
change local to home and foreign to away
}
