% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparisons.check_model_anova.R
\name{mean_comparisons.check_model_anova}
\alias{mean_comparisons.check_model_anova}
\title{Get mean comparisons from \code{\link{check_model.fit_model_anova}} object}
\usage{
\method{mean_comparisons}{check_model_anova}(x, alpha = 0.05, p.adj = "none", ...)
}
\arguments{
\item{x}{outputs from \code{\link{check_model.fit_model_anova}}}

\item{alpha}{level of type one error. 0.05 (5\%) by default}

\item{p.adj}{For all except type = 2. 
NULL for no adjustement of the type one error. 
p.adj can "holm", "hochberg", "bonferroni", "BH", "BY" or "fdr"
p-adj = "none" is t-student.
See p.adjust() for more details.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list of four elements : 
  \itemize{
   \item info : a list with variable
  }
}
\description{
\code{mean_comparisons} performs mean comparisons from object coming from \code{\link{check_model.fit_model_anova}}
}
\details{
S3 method.
See in the book for more details \href{https://priviere.github.io/PPBstats_book/intro-agro.html#section-freq}{here}
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{plot.PPBstats}}
 \item \code{\link{plot.mean_comparisons_model_anova}}
}
}
\author{
Pierre Riviere
}
