% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparisons.check_model_variance_intra.R
\name{mean_comparisons.check_model_bh_variance_intra}
\alias{mean_comparisons.check_model_bh_variance_intra}
\title{Get mean comparisons from \code{\link{check_model.fit_model_bh_variance_intra}} object}
\usage{
\method{mean_comparisons}{check_model_bh_variance_intra}(x, parameter,
  alpha = 0.05, type = 1, get.at.least.X.groups = 2, precision = 5e-04,
  threshold = 1, p.adj = "soft.bonf", ...)
}
\arguments{
\item{x}{outputs from \code{\link{check_model.fit_model_bh_variance_intra}}}

\item{parameter}{parameter on which the mean comparison is done. 
The possible values are "mu" and "sigma"}

\item{alpha}{level of type one error. 0.05 (5\%) by default}

\item{type}{type of comparisons
\itemize{
 \item type = 1 for comparison two by two
 \item type = 2 for comparison to a specific threshold
}}

\item{get.at.least.X.groups}{For type = 1. 
If there are only one group with alpha, the minimum number of groups wanted with a higher type one error (i.e. lower confidence). 
If NULL, nothing is done.}

\item{precision}{For type = 1. The precision of the alpha with the correspondong groups from get.at.least.X.groups. The smaller the better, but the smaller the more time consuming due to computing matters}

\item{threshold}{For type = 2. The threshold to which a parameter is different}

\item{p.adj}{For all except type = 2. 
NULL for no adjustement of the type one error. 
p.adj can be "soft.bonf". 

p.adj = "soft.bonf" for a soft bonferonni correction to take into account multiple comparisons (alpha / nb of parameters)..
The comparisons is based on the probability of having a common distribution for each pair of parameter.
When there is only one group with the value of alpha, the function (via \code{get.at.least.X.groups argument}) returns at least X groups with a new value of alpha.}

\item{...}{further arguments passed to or from other methods#'}
}
\value{
A list of one elements : 
  \itemize{
   \item data_mean_comparisons a list with as many elements as environment.
   Each element of the list is composed of two elements:
   \itemize{
    \item mean.comparisons: a dataframe with the following columns : parameter, median, groups, number of groups, alpha (type one error), alpha.correction (correction used), entry, environment, location and year.
    \item Mpvalue : a square matrix with pvalue computed for each pair of parameter.
   }
  }
}
\description{
\code{mean_comparisons} performs mean comparisons from object coming from \code{\link{check_model.fit_model_bh_variance_intra}}
}
\details{
S3 method.
For more details, see in the book : https://priviere.github.io/PPBstats_book/intro-agro.html#section-bayes
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{plot.PPBstats}}
}
}
\author{
Pierre Riviere
}
