% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparisons.check_model_home_away.R
\name{mean_comparisons.check_model_home_away}
\alias{mean_comparisons.check_model_home_away}
\title{Get mean comparisons from \code{\link{check_model.fit_model_home_away}} object}
\usage{
\method{mean_comparisons}{check_model_home_away}(x, alpha = 0.05,
  p.adj = "none", ...)
}
\arguments{
\item{x}{outputs from \code{\link{check_model.fit_model_home_away}}}

\item{alpha}{level of type one error. 0.05 (5\%) by default}

\item{p.adj}{NULL for no adjustement of the type one error. 
p.adj can be "tukey".}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list of four elements : 
  \itemize{
   \item info : a list with variable and home_away analysis
   \item data_ggplot_LSDbarplot_version:germplasm
   \item data_ggplot_LSDbarplot_germplasm
   \item data_ggplot_LSDbarplot_location
   \item data_ggplot_LSDbarplot_year
  }
}
\description{
\code{mean_comparisons} performs mean comparisons from object coming from \code{\link{check_model.fit_model_home_away}}
}
\details{
S3 method.
Mean comparisons based on LSmeans.
See in the book for mo arere details \href{https://priviere.github.io/PPBstats_book/intro-agro.html#section-freq}{here}
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{plot.PPBstats}}
 \item \code{\link{plot.mean_comparisons_model_home_away}}
}
}
\author{
Pierre Riviere
}
