% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_comparisons_predict_the_past_model_2.R
\name{mean_comparisons.predict_the_past_model_2}
\alias{mean_comparisons.predict_the_past_model_2}
\title{Get mean comparisons from predict_the_past_model_2}
\usage{
\method{mean_comparisons}{predict_the_past_model_2}(x, alpha = 0.05,
  type = 1, get.at.least.X.groups = 2, precision = 5e-04, threshold = 1,
  p.adj = "soft.bonf")
}
\arguments{
\item{x}{}

\item{alpha}{}

\item{type}{}

\item{get.at.least.X.groups}{}

\item{precision}{}

\item{threshold}{}

\item{p.adj}{}
}
\value{
See \code{\link{mean_comparisons}}
}
\description{
\code{mean_comparisons_predict_the_past_model_2} performs mean comparisons from object coming from \code{\link{predict_the_past_model_2}}
See \code{\link{mean_comparisons}} for more information.
}
\details{
See \code{\link{mean_comparisons}}
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{get_ggplot}}
}
}

