% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparisons.R
\name{mean_comparisons}
\alias{mean_comparisons}
\title{Get mean comparisons from \code{\link{check_model}} object}
\usage{
mean_comparisons(x, ...)
}
\arguments{
\item{x}{outputs from 
\itemize{
 \item \code{\link{check_model.fit_model_bh_GxE}}
 \item \code{\link{check_model.fit_model_bh_intra_location}}
 \item \code{\link{check_model.fit_model_bh_variance_intra}}
 \item \code{\link{check_model.fit_model_GxE}}
 \item \code{\link{check_model.fit_model_hedonic}}
 \item \code{\link{check_model.fit_model_spatial}}
 \item \code{\link{predict_the_past_model_bh_GxE}}
}}

\item{...}{See details}
}
\value{
It returns a list of elements ready to be plotted
}
\description{
\code{mean_comparisons} performs mean comparisons from object coming from \code{\link{check_model}} or \code{\link{predict_the_past_model_bh_GxE}}
}
\details{
S3 method.
See for more details :
\itemize{
 \item \code{\link{mean_comparisons.check_model_GxE}}
 \item \code{\link{mean_comparisons.check_model_bh_intra_location}}
 \item \code{\link{mean_comparisons.check_model_bh_GxE}}
 \item \code{\link{mean_comparisons.check_model_hedonic}}
 \item \code{\link{mean_comparisons.check_model_spatial}}
 \item \code{\link{mean_comparisons.check_model_bh_variance_intra}}
}
}
\seealso{
\itemize{
 \item \code{\link{check_model}}
 \code{\link{predict_the_past_model_bh_GxE}}
 \code{\link{plot.PPBstats}}
}
}
\author{
Pierre Riviere
}
