% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{mean_comparisons_freq_anova}
\alias{mean_comparisons_freq_anova}
\title{mean comparisons for frequentist analysis}
\usage{
mean_comparisons_freq_anova(model, variable, alpha = 0.05,
  p.adj = "none", info = NULL, vec_fac = c("germplasm", "location",
  "year"))
}
\arguments{
\item{model}{anova model}

\item{variable}{variable}

\item{alpha}{alpha}

\item{p.adj}{p.adj}

\item{info}{info from mean_comparisons}

\item{vec_fac}{vec_fac}
}
\description{
mean comparisons for frequentist analysis
}
