% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_comparisons_GxE.R
\name{mean_comparisons_GxE}
\alias{mean_comparisons_GxE}
\title{Get mean comparisons from object coming from check_model GxE}
\usage{
mean_comparisons_GxE(out_check_model_GxE, alpha = 0.05, p.adj = "none")
}
\arguments{
\item{out_check_model_GxE}{}

\item{alpha}{}

\item{p.adj}{}
}
\value{
See mean_comparisons
}
\description{
\code{mean_comparisons_GxE} performs mean comparisons from object coming from check_model GxE
See \code{mean_comparisons} for more information.
}
\details{
See mean_comparisons
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{get_ggplot}}
}
}

