% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_2.R
\name{model_2}
\alias{model_2}
\title{Run model 2 to get main germplasm, environment and sensitivity effects over the network.}
\usage{
model_2(data, variable, nb_iterations = 1e+05, thin = 10,
  return.alpha = TRUE, return.sigma_alpha = TRUE, return.beta = TRUE,
  return.sigma_beta = TRUE, return.theta = TRUE,
  return.sigma_theta = TRUE, return.epsilon = FALSE,
  return.sigma_epsilon = TRUE, return.DIC = FALSE)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{The variable on which runs the model}

\item{nb_iterations}{Number of iterations of the MCMC}

\item{thin}{thinning interval to reduce autocorrelations between samples of the MCMC}

\item{return.alpha}{Return the value for each germplasm main effect (alpha_i)}

\item{return.sigma_alpha}{Return the value of the variance of the distribution where the alpha_i come from}

\item{return.beta}{Return the value for each sensitivity to environments (beta_i)}

\item{return.sigma_beta}{Return the value of the variance of the distribution where the beta_i come from}

\item{return.theta}{Return the value for each environment main effect (theta_j)}

\item{return.sigma_theta}{Return the value of the variance of the distribution where the theta_j come from}

\item{return.epsilon}{Return the value of the residuals of the model (epsilon_ij)}

\item{return.sigma_epsilon}{Return the value of the variance of the distribution where the epsilon_ij come from}

\item{return.DIC}{Return the DIC value of the model. See details for more informations.}
}
\value{
The function returns a list with

\itemize{
\item "data.presence.abscence.matrix": a matrix germplasm x environment with the number of occurence in the data
\item "model.presence.abscence.matrix": a matrix germplasm x environment with the number of occurence in the data used for the model (i.e. with at least two germplasm by environments.)
\item "germplasm.not.used": the vector of germplasms not used in the analysis because they were not on at least two environments. If NULL, all the germplasms were used in the analysis.
\item "MCMC": a list with the two MCMC chains (mcmc object) from the model
\item "epsilon": a vector with the median value of the epsilon_ijk
\item "DIC": the DIC value of the model
}
}
\description{
\code{model_2} runs model 2 known as the Finlay Wilkinson Hierarchical model to get main germplasm, environment and sensitivity effects over the network
}
\details{
Model 2 estimates germplasm (alpha_i), environment (theta_j) and sensitivity to interaction (beta_i) effects.
An environment is a combinaison of a location and a year.

The different effects are taken in different distributions of respective variances sigma_alpha, sigma_theta and sigma_beta.
This model takes into acount all the information on the network in order to cope with the high disequilibrium in the dataset (i.e. high percentage of missing GxE combinaisons on the network).

First, the additive model is done. This model gives intitial values of some parameters of the Hierarchical Finlay Wilkinson model which is done next.

The model is run on data set where germplasms are on at least two environments

More informations can be found in the vignette.

For DIC value, see ?\code{dic.samples} from the \code{rjags} package for more information.
}
\author{
Pierre Riviere for R code and Olivier David for JAGS code
}
\references{
P. Riviere, J.C. Dawson, I. Goldringer, and O. David. Hierarchical multiplicative modeling of genotype x environment interaction for flexible experiments in decentralized participatory plant breeding. In prep, 2015.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model_model_2}}
\item \code{\link{cross_validation_model_2}},
\item \code{\link{predict_the_past_model_2}}
}
}

