% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_anova.R
\name{model_anova}
\alias{model_anova}
\title{Run classic anova model}
\usage{
model_anova(data, variable)
}
\arguments{
\item{data}{The data frame on which the model is run. It should come from \code{\link{format_data_PPBstats.data_agro}}}

\item{variable}{variable to analyse}
}
\value{
The function returns a list with three elements :
\itemize{
 \item info : a list with variable and gxe_analysis
 \item ANOVA a list with five elements:
  \itemize{
   \item model
   \item anova_model
   \item germplasm_effects a list of two elements:
    \itemize{
     \item effects
     \item intra_variance
    }
  }
}
}
\description{
\code{model_anova} runs classic anova model
}
\details{
More information can be found in the book regarding \href{https://priviere.github.io/PPBstats_book/family-1.html#classic-anova}{classic anova}.
}
\seealso{
\itemize{
\item \code{\link{check_model}},
\item \code{\link{check_model.fit_model_anova}}
}
}
\author{
Pierre Riviere
}
