% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bh_variance_intra.R
\name{model_bh_variance_intra}
\alias{model_bh_variance_intra}
\title{Run Hierarchical Bayesian variance-intra model}
\usage{
model_bh_variance_intra(data, variable, nb_iterations = 1e+05, thin = 10,
  return.mu = TRUE, return.sigma = TRUE, return.epsilon = FALSE,
  return.DIC = FALSE)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{The variable on which runs the model}

\item{nb_iterations}{Number of iterations of the MCMC}

\item{thin}{thinning interval to reduce autocorrelations between samples of the MCMC}

\item{return.mu}{Return the value for each entry in each environment and each plot (mu_ijk)}

\item{return.sigma}{Return the value for each intra-population variance in each environment (sigma_ij)}

\item{return.epsilon}{Return the value of all residuals in each environment on each plot (epsilon_ijk)}

\item{return.DIC}{Return the DIC value of the model. See details for more information.}
}
\value{
The function returns a list with 

\itemize{
\item "data.model_bh_variance_intra": the dataframe used to run mode variance_intra
\item "MCMC": a list with the two MCMC chains (mcmc object)
\item "DIC": the DIC value of the model
}
}
\description{
\code{model_bh_variance_intra} runs Hierarchical Bayesian variance-intra model to get intra-population variance on each environment of the network. See details for more information.
}
\details{
Model on intra-population variance estimates entry effects (mu_ijk) and within-population variance (sigma_ij) on each environment. 
An environment is a combinaison of a location and a year.

The variance are taken in an inverse Gamma distribution of parameters 10^-6. 
More information can be found in the vignette.

For DIC value, see ?\code{dic.samples} from the \code{rjags} package for more information.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.model_bh_variance_intra}}
}
}
\author{
Pierre Riviere and Gaelle van Frank for R code; Olivier David for JAGS code
}
