% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_hedonic.R
\name{model_hedonic}
\alias{model_hedonic}
\title{Run hedonic analysis}
\usage{
model_hedonic(data)
}
\arguments{
\item{data}{data is a data frame with the following columns: sample, juges, note, descriptors, germplasm, location. 
The descriptors must be separated by ";". Any other column can be added as supplementary variables.}
}
\value{
A list with the anova and the CA object
}
\description{
\code{model_hedonic} runs hedonic analysis
}
\details{
An anova is run on "note".
A Correspondence Analysis is run with FactoMineR::CA
}
