% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_local_foreign.R
\name{model_local_foreign}
\alias{model_local_foreign}
\title{Run Local Foreign model}
\usage{
model_local_foreign(data, data_version, variable)
}
\arguments{
\item{data}{The data frame on which the model is run.
It should come from \code{\link{format_data_PPBstats.data_agro}}}

\item{data_version}{It should come from \code{\link{format_data_PPBstats.data_agro_version}} with local and foreign in version}

\item{variable}{variable to analyse}
}
\value{
The function returns a list with three elements :
\itemize{
 \item info : a list with variable
 \item ANOVA a list with two elements:
  \itemize{
   \item model
   \item anova_model
  }
}
}
\description{
\code{model_local_foreign} runs local foreign model
}
\details{
Find details in the book \href{https://priviere.github.io/PPBstats_book/family-4.html#family-4}{here}.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_local_foreign}}
}
}
\author{
Pierre Riviere and Gaelle Van Frank
}
