% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_napping.R
\name{model_napping}
\alias{model_napping}
\title{Run napping analysis}
\usage{
model_napping(data)
}
\arguments{
\item{data}{data frame coming from \code{\link{format_data_PPBstats.data_organo_napping}}}
}
\value{
A list with two elements : the MFA object and the data
}
\description{
\code{model_napping} runs napping analysis
}
\details{
The Multiple Factor Analysis is run with FactoMineR::MFA

More information can be found in the book : https://priviere.github.io/PPBstats_book/napping.html
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_napping}}
}
}
\author{
Pierre Riviere and Camille Vindras
}
