% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{model_spatial}
\alias{model_spatial}
\title{Run spatial row and column model}
\usage{
model_spatial(data, variable, genotype.as.random = TRUE)
}
\arguments{
\item{data}{The data frame on which the model is run. 
It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{variable to analyse}

\item{genotype.as.random}{TRUE of FALSE. TRUE by default.}
}
\value{
The function returns a list with two elements :
\itemize{
 \item info : a list with variable
 \item model : 
  \itemize{
   \item model : the output from SpATS function
   \item summary : summary of the model
   \item germplasm_effects : germplasm effects, BLUE or BLUP depending if it is fixed or random effect
   \item var_res : variance of the residuals
  }
}
}
\description{
\code{model_spatial} runs spatial row and column model based on the SpATS package
}
\details{
The model is run with the SpATS function of package SpATS. 
See ?SpATS for more information.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
}
}
\author{
Pierre Riviere
}
