% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_variance_intra.R
\name{model_variance_intra}
\alias{model_variance_intra}
\title{Run model variance_intra}
\usage{
model_variance_intra(data, variable, nb_iterations = 1e+05, thin = 10,
  return.mu = TRUE, return.sigma = TRUE, return.DIC = FALSE)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{The variable on which runs the model}

\item{nb_iterations}{Number of iterations of the MCMC}

\item{thin}{thinning interval to reduce autocorrelations between samples of the MCMC}

\item{return.mu}{Return the value for each entry in each environment (mu_ij)}

\item{return.sigma}{Return the value for each within-environment variance (sigma_j)}

\item{return.DIC}{Return the DIC value of the model. See details for more information.}
}
\value{
The function returns a list with 

\itemize{
\item "data.model_variance_intra": the dataframe used to run mode variance_intra
\item "vec_env_RF": a vector with the environments as regional farms (i.e. with at least two blocks)
\item "vec_env_SF": a vector with the environments as satellite farms (i.e. with one block)
\item "MCMC": a list with the two MCMC chains (mcmc object)
\item "DIC": the DIC value of the model
}
}
\description{
\code{model_variance_intra} runs model variance_intra
}
\details{
TO DO, see model 1 to get inspiration
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.model_variance_intra}}
}
}
\author{
Pierre Riviere and Gaelle Van Frank for R code; Olivier David for JAGS code
}
