% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_variance_intra.R
\name{model_varintra}
\alias{model_varintra}
\title{Run model to analyse intra-population variance}
\usage{
model_varintra(data, variable, nb_iterations = 1e+05, thin = 10,
  return.mu = TRUE, return.sigma = TRUE, return.DIC = FALSE)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}
}
\description{
\code
}
