% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{mutlivariate}
\alias{mutlivariate}
\title{Run multivariate analysis with function from FactoMineR}
\usage{
mutlivariate(data, vec_variables, FUN)
}
\arguments{
\item{data}{The data frame. See details.}

\item{vec_variables}{vector of variables to use in the analysis}

\item{FUN}{function to apply : PCA}
}
\value{
The function returns the FactoMineR object corresponding to FUN
}
\description{
\code{mutlivariate} runs multivariate analysis with function from FactoMineR
}
\details{
This function is useful to settle the right data set and quali.sup
}
\author{
Pierre Riviere
}
