% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napping.R
\name{napping}
\alias{napping}
\title{Run napping analysis}
\usage{
napping(data)
}
\arguments{
\item{data}{data frame coming from format_data_PPBstats().
It has the following columns: sample, juges, X, Y, descriptors. 
The descriptors must be separated by ";"}
}
\value{
A list of ? elements being the MFA and obecjt and its graph :
\itemize{
 \item out_MFA
 \item 
 \item 
}
}
\description{
\code{napping} runs napping analysis
}
\details{
The Multiple Factor Analysis is run with FactoMineR::MFA on d_freq
}
\author{
Pierre Riviere and Camille Vindras
}
