% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_groups.R
\name{parameter_groups}
\alias{parameter_groups}
\title{Get groups of parameters based on multivariate analysis}
\usage{
parameter_groups(list_out_check_model, parameter)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{check_model}}

\item{parameter}{The parameter on which to get the groups}
}
\value{
The function returns a list of two elements:
\itemize{
 \item obj.pca : the PCA object from FactoMineR::PCA
 \item clust, a list of four elements:
  \itemize{
   \item data : the matrix with variables in column and effect in row, from wich have been computed the PCA
   \item nb.k : number of cluster
   \item hc : hcluster
   \item km : kmeans
  }
}
}
\description{
\code{parameter_groups} gets groups of parameters based on multivariate analysis
}
\details{
The function run a PCA on a matrix containing the value of each parameter for each variable.

Clusters are done based on Average silhouette method for k-means clustering as explained here \url{http://www.sthda.com/english/wiki/determining-the-optimal-number-of-clusters-3-must-known-methods-unsupervised-machine-learning#r-codes-1}

The hcluster are computed with the function hclust(dist())

The kmeans are computed with the function kmeans()
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
 \item \code{\link{check_model}},
 \item \code{\link{check_model_model_2}},
 \item \code{\link{check_model_GxE}},
 \item \code{\link{parameter_groups_GxE}},
 \item \code{\link{parameter_groups_model_2}},
 \item \code{\link{get_ggplot}}
}
}

