% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_groups_GxE.R
\name{parameter_groups_GxE}
\alias{parameter_groups_GxE}
\title{Get matrix with variables in column and effect in row from \code{check_model_GxE}}
\usage{
parameter_groups_GxE(list_out_check_model_GxE, parameter)
}
\arguments{
\item{list_out_check_model_GxE}{A list whose elements are output from \code{check_model_GxE}}

\item{parameter}{parameter on which to get the matrix}
}
\value{
The function returns a data frame with variables in column and effect of paramters in row
}
\description{
\code{parameter_groups_GxE} gets matrix with variables in column and effect in row
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
 \item \code{\link{parameter_groups}},
 \item \code{\link{parameter_groups_model_2}},
}
}

