% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_groups_model_2.R
\name{parameter_groups_model_2}
\alias{parameter_groups_model_2}
\title{Get matrix with variables in column and effect in row from check_model_model_2}
\usage{
parameter_groups_model_2(list_out_check_model_model_2, parameter)
}
\arguments{
\item{list_out_check_model_model_2}{A list whose elements are output from \code{\link{check_model_model_2}}}

\item{parameter}{parameter on which to get the matrix}
}
\value{
The function returns a data frame with variables in column and effect of paramters in row
}
\description{
\code{parameter_groups_model_2} gets matrix with variables in column and effect in row
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{parameter_groups}}
}

