% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_groups_model_anova.R
\name{parameter_groups_model_anova}
\alias{parameter_groups_model_anova}
\title{Get matrix with variables in column and effects in row from \code{\link{check_model.fit_model_anova}}}
\usage{
parameter_groups_model_anova(list_out_check_model_anova, parameter)
}
\arguments{
\item{list_out_check_model_anova}{A list whose elements are output from \code{\link{check_model.fit_model_anova}}}

\item{parameter}{parameter on which to get the matrix: germplasm, location or var_intra_germplasm}
}
\value{
The function returns a data frame with variables in column and effect of paramters in row
}
\description{
\code{parameter_groups_model_anova} gets matrix with variables in column and effects in row
}
\seealso{
\code{\link{parameter_groups}}
}
\author{
Pierre Riviere
}
