% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_groups_model_spatial.R
\name{parameter_groups_model_spatial}
\alias{parameter_groups_model_spatial}
\title{Get matrix with variables in column and effects in row from \code{\link{check_model.fit_model_spatial}}}
\usage{
parameter_groups_model_spatial(list_out_check_model_spatial, parameter)
}
\arguments{
\item{list_out_check_model_spatial}{A list whose elements are output from \code{\link{check_model.fit_model_spatial}}}

\item{parameter}{parameter on which to get the matrix: germplasm}
}
\value{
The function returns a data frame with variables in column and effect of paramters in row
}
\description{
\code{parameter_groups_model_spatial} gets matrix with variables in column and effects in row
}
\seealso{
\code{\link{parameter_groups}}
}
\author{
Pierre Riviere
}
