% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plan_experiment.R
\name{plan_experiment}
\alias{plan_experiment}
\title{Provides experimental plan for several situations}
\usage{
plan_experiment(expe.type, nb.entries, nb.controls)
}
\arguments{
\item{expe.type}{The type of experiment to settle "satellite-farm", "regional-farm", "row-columns", "alpha-design", "fully-repicated"}

\item{nb.entries}{Number of entries}

\item{nb.controls}{Number of controls. This is useful only for expe.type "satellite-farm" and "regional-farm"}

\item{nb.blocks}{Number of blocks}

\item{nb.cols}{Number of columns}

\item{nb.rows}{Number of rows}
}
\value{
The function returns a list with
\itemize{
 \item A data frame with X, Y and block
 \item A picture of the experimental plan, which is also exported to csv files
 }
}
\description{
\code{plan_experiment} provides experimental plan for several situations
}
\author{
Pierre Riviere
}

