% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.biplot_hedonic.R
\name{plot.biplot_hedonic}
\alias{plot.biplot_hedonic}
\title{Get ggplot to visualize output from \code{\link{biplot_data.check_model_hedonic}}}
\usage{
\method{plot}{biplot_hedonic}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{biplot_data.check_model_hedonic}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
It returns a list with 3 elements,  which are outputs from CA analysis:
  \itemize{
   \item ind
   \item var
   \item axes
  }
}
\description{
\code{plot.biplot_hedonic} returns ggplot to visualize outputs from \code{\link{biplot_data.check_model_hedonic}}
}
\details{
S3 method.
The plot are done with the factoextra package
}
\seealso{
\itemize{
 \item \code{biplot_data}
}
}
\author{
Pierre Riviere
}
