% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_check_model_model_1.R
\name{plot.check_model_1}
\alias{plot.check_model_1}
\title{Get ggplot from check_model_model_1}
\usage{
\method{plot}{check_model_1}(x, nb_parameters_per_plot = 8)
}
\arguments{
\item{x}{outputs from \code{\link{check_model_model_1}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_check_model_model_1} returns ggplot from \code{\link{check_model_model_1}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}}, 
\item \code{\link{check_model_model_1}}
}
}
