% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_model_bh_variance_intra.R
\name{plot.check_model_bh_variance_intra}
\alias{plot.check_model_bh_variance_intra}
\title{Get ggplot to visualize output from \code{\link{check_model.fit_model_bh_variance_intra}}}
\usage{
plot.check_model_bh_variance_intra(x, nb_parameters_per_plot = 10, ...)
}
\arguments{
\item{x}{Output from \code{\link{check_model.fit_model_bh_variance_intra}}}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
\itemize{
 \item mu_ijk : distribution of each mu_ijk. 
 There are as many graph as needed with \code{nb_parameters_per_plot} alpha_i per graph.
 \item sigma_ij : distribution of each sigma_ij. 
 There are as many graph as needed with \code{nb_parameters_per_plot} alpha_i per graph.
 \item mcmc_not_converge_traceplot_density : a list with the plots of trace and density to check the convergence of the two MCMC only for chains that are not converging thanks to the Gelman-Rubin test. 
 If all the chains converge, it is NULL.
 }
}
\description{
\code{plot.check_model_bh_variance_intra} returns ggplot to visualize outputs from \code{\link{check_model.fit_model_bh_variance_intra}}
}
\details{
S3 method.

For mcmc_not_converge_traceplot_density : If you wish exhaustive information, look at \code{ggmcmc::ggmcmc} with \code{ggmcmc(out_model$MCMC)}. 
But be careful with the size of your MCMC output which are often too big to be performed in R.

See example in the book: https://priviere.github.io/PPBstats_book/family-4.html#variance-intra
}
\seealso{
\code{\link{check_model.fit_model_bh_variance_intra}}
}
\author{
Gaelle Van Frank and Pierre Riviere
}
