% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_model_hedonic.R
\name{plot.check_model_hedonic}
\alias{plot.check_model_hedonic}
\title{Get ggplot to visualize output from \code{\link{check_model.fit_model_hedonic}}}
\usage{
plot.check_model_hedonic(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{check_model.fit_model_hedonic}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\itemize{
 \item residuals
 \itemize{
  \item histogram : histogram with the distribution of the residuals
  \item qqplot
  \item points
  }
 \item variability_repartition : pie with repartition of SumSq for each factor
 \item variance_intra_germplasm : repartition of the residuals for each germplasm (see Details for more information)
 \item CA_composante_variance : variance caught by each dimension of the CA
 \item PCA_composante_variance : variance caught by each dimension of the PCA previous to the HCPC
 }
}
\description{
\code{plot.check_model_hedonic} returns ggplot to visualize outputs from \code{\link{check_model.fit_model_hedonic}}
}
\details{
S3 method.

variance_intra_germplasm display the repartition of the residuals for each germplasm which represent the person assessor variation plus the intra-germplasm variance

See example in the book : https://priviere.github.io/PPBstats_book/hedonic.html#check-and-visualize-model-outputs-6
}
\seealso{
\code{\link{check_model.fit_model_hedonic}}
}
\author{
Pierre Riviere
}
