% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_model_napping.R
\name{plot.check_model_napping}
\alias{plot.check_model_napping}
\title{Get ggplot to visualize output from \code{\link{check_model.fit_model_napping}}}
\usage{
\method{plot}{check_model_napping}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{check_model.fit_model_napping}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
A plot with the variance caught by the dimension of the MFA
}
\description{
\code{plot.check_model_napping} returns ggplot to visualize outputs from \code{\link{check_model.fit_model_napping}}
}
\details{
S3 method.
See example in the book : https://priviere.github.io/PPBstats_book/napping.html#check-and-visualize-model-outputs-7
}
\seealso{
\code{\link{check_model.fit_model_napping}}
}
\author{
Pierre Riviere
}
