% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_cross_validation_model_2.R
\name{plot.cross_validation_model_2}
\alias{plot.cross_validation_model_2}
\title{Get ggplot from cross_validation_model_2}
\usage{
\method{plot}{cross_validation_model_2}(x)
}
\arguments{
\item{x}{outputs from \code{\link{cross_validation_model_2}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_cross_validation_model_2} returns ggplot from \code{\link{cross_validation_model_2}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}}, 
\item \code{\link{cross_validation_model_2}}
}
}
