% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cross_validation_model_bh_GxE.R
\name{plot.cross_validation_model_bh_GxE}
\alias{plot.cross_validation_model_bh_GxE}
\title{Get ggplot to visualize output from \code{\link{cross_validation_model_bh_GxE}}}
\usage{
plot.cross_validation_model_bh_GxE(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{cross_validation_model_bh_GxE}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\itemize{
   \item plot : plot estimated.value = f(observed.value). 
   The probability mean = 0 is display (see Details for more information).
   \item regression : output of the model observed.value = a x estimated.value + b
  }
}
\description{
\code{plot.cross_validation_model_bh_GxE} returns ggplot to visualize outputs from \code{\link{cross_validation_model_bh_GxE}}
}
\details{
S3 method.
For plot : plot estimated.value = f(observed.value). 
The probability mean = 0 is coming from a \code{t.test} performed with the null hypothesis H0: the bias of estimated values in relation to real values = 0.

See example in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-2
}
\seealso{
\code{\link{cross_validation_model_bh_GxE}}
}
\author{
Pierre Riviere
}
