% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_agro.R
\name{plot.data_agro}
\alias{plot.data_agro}
\title{Plot agro object from format_data_PPBstats()}
\usage{
\method{plot}{data_agro}(data, plot_type = c("pam", "histogramm", "barplot",
  "boxplot", "interaction", "biplot", "radar", "raster"), x_axis = NULL,
  in_col = NULL, vec_variables = NULL, nb_parameters_per_plot_x_axis = 5,
  nb_parameters_per_plot_in_col = 5, labels_on = NULL, labels_size = 4)
}
\arguments{
\item{data}{The data frame. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{plot_type}{the type of plot you wish. It can be :
\itemize{
 \item "pam" for presence abscence matrix that represent the combinaison of germplasm x location
 \item "histogramm"
 \item "barplot", where sd error are displayed
 \item "boxplot"
 \item "interaction"
 \item "biplot"
 \item "radar"
 \item "raster"
}}

\item{x_axis}{factor displayed on the x.axis of a plot}

\item{in_col}{factor displayed in color of a plot}

\item{vec_variables}{vector of variables to describe}

\item{nb_parameters_per_plot_x_axis}{the number of parameters per plot on x_axis arguments}

\item{nb_parameters_per_plot_in_col}{the number of parameters per plot for in_col arguments}

\item{labels_on}{factor to display for plot_type = "biplot"}

\item{labels_size}{size of the label for plot_type = "biplot" and "radar"}
}
\value{
\itemize{
 \item For plot_type "histogramm", "barplot", "boxplot" or "interaction",
 the function returns a list with ggplot objects for each variable of vec_variables.
 \item For plot_type "biplot",
 the function returns a list with ggplot objects for each pairs of variables of vec_variables. 
 \item For plot_type "radar",
 the function returns a list with ggplot objects with all variables of vec_variables. 
 \item For plot_type "raster"
 the function returns a list with ggplot objects with all variables of vec_variables. 
}
Each list is divided in several lists according to values 
of nb_parameters_per_plot_x_axis and nb_parameters_per_plot_in_col.
}
\description{
\code{plot.data_agro} gets ggplot to describe the data
}
\author{
Pierre Riviere
}
