% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_agro_HA.R
\name{plot.data_agro_HA}
\alias{plot.data_agro_HA}
\title{Plot agro object from format_data_PPBstats.data_agro_HA()}
\usage{
\method{plot}{data_agro_HA}(x, plot_type = "boxplot",
  vec_variables = NULL, ...)
}
\arguments{
\item{x}{The data frame. It should come from \code{\link{format_data_PPBstats.data_agro_HA}}}

\item{plot_type}{the type of plot you wish. It can be :
\itemize{
 \item "barplot", where sd error are displayed
 \item "boxplot"
}}

\item{vec_variables}{vector of variables to display}
}
\value{
The function returns a list with ggplot objects for each variable of vec_variables divided into three elements:
\itemize{
 \item home_away_merged, i.e. a single plot with version for all germplasm merged
 \item home_away_merged_per_germplasm, i.e. a single plot with version for each germplasm
 \item home_away_per_germplasm, i.e. a list of plots for each germplasm with all version separated
}

When argument mean_comparison is not NULL, it returns a plot with stars of significant differences.
}
\description{
\code{plot.data_agro_HA} gets ggplot to describe the data
}
\details{
S3 method.
See the book for more details \href{https://priviere.github.io/PPBstats_book/home-away.html}{here}.
}
\seealso{
\itemize{
 \item \code{\link{format_data_PPBstats}}
 \item \code{\link{format_data_PPBstats.data_agro}}
}
}
\author{
Pierre Riviere
}
