% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_agro_LF.R
\name{plot.data_agro_LF}
\alias{plot.data_agro_LF}
\title{Plot agro object from format_data_PPBstats.data_agro_LF()}
\usage{
\method{plot}{data_agro_LF}(x, plot_type = "boxplot",
  vec_variables = NULL, ...)
}
\arguments{
\item{x}{The data frame. It should come from \code{\link{format_data_PPBstats.data_agro_LF}}}

\item{plot_type}{the type of plot you wish. It can be :
\itemize{
 \item "barplot", where sd error are displayed
 \item "boxplot"
}}

\item{vec_variables}{vector of variables to display}
}
\value{
The function returns a list with ggplot objects for each variable of vec_variables divided into three elements:
\itemize{
 \item local_foreign_merged, i.e. a single plot with version for all location merged
 \item local_foreign_merged_per_location, i.e. a single plot with version for each location
 \item local_foreign_per_location, i.e. a list of plots for each location with all version separated
}

When argument mean_comparison is not NULL, it returns a plot with stars of significant differences.
}
\description{
\code{plot.data_agro_LF} gets ggplot to describe the data
}
\details{
S3 method.
See the book for more details \href{https://priviere.github.io/PPBstats_book/local-foreign.html}{here}.
}
\seealso{
\itemize{
 \item \code{\link{format_data_PPBstats}}
 \item \code{\link{format_data_PPBstats.data_agro}}
}
}
\author{
Pierre Riviere
}
