% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_agro_SR.R
\name{plot.data_agro_SR}
\alias{plot.data_agro_SR}
\title{Plot agro object from format_data_PPBstats.data_agro_SR()}
\usage{
plot.data_agro_SR(x, mean_comparisons = NULL, plot_type = c("barplot",
  "boxplot", "interaction"), x_axis = NULL, in_col = NULL,
  vec_variables = NULL, nb_parameters_per_plot_x_axis = 5,
  nb_parameters_per_plot_in_col = 5, ...)
}
\arguments{
\item{x}{The data frame. It should come from \code{\link{format_data_PPBstats.data_agro_SR}}}

\item{mean_comparisons}{Output from \code{\link{mean_comparisons}}.}

\item{plot_type}{the type of plot you wish. It can be :
\itemize{
 \item "barplot", where sd error are displayed
 \item "boxplot"
 \item "interaction"
}}

\item{vec_variables}{vector of variables to display}

\item{nb_parameters_per_plot_x_axis}{the number of parameters per plot on x_axis arguments}

\item{nb_parameters_per_plot_in_col}{the number of parameters per plot for in_col arguments}

\item{...}{further arguments passed to or from other methods}
}
\value{
The function returns a list with ggplot objects for each variable of vec_variables divided then for each element expe_id.
When argument mean_comparison is not NULL, it returns a plot with stars of significant differences.
}
\description{
\code{plot.data_agro_SR} gets ggplot to describe the data
}
\details{
S3 method.
See the book for more details \href{https://priviere.github.io/PPBstats_book/response-to-selection.html}{here}.
}
\seealso{
\itemize{
 \item \code{\link{format_data_PPBstats.data_agro_SR}}
}
}
\author{
Pierre Riviere
}
