% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_network.R
\name{plot.data_network}
\alias{plot.data_network}
\title{Plot network object from format_data_PPBstats()}
\usage{
\method{plot}{data_network}(x, data_to_pie = NULL, variable = NULL,
  pie_size = 0.2, plot_type = "network", in_col = "location",
  labels_on = FALSE, labels_size = 4, organize_sl = FALSE,
  x_axis = "germplasm", nb_parameters_per_plot_x_axis = 5,
  nb_parameters_per_plot_in_col = 5, vec_variables = NULL, zoom = 6,
  ...)
}
\arguments{
\item{x}{output from \code{\link{format_data_PPBstats.data_network}}}

\item{data_to_pie}{output from format_data_PPBstats with data_type = "data_agro"}

\item{variable}{when data_to_pie is not NULL, variable to plot in a pie}

\item{pie_size}{when data_to_pie is not NULL, size of the pie}

\item{plot_type}{"network", "barplot" or "map"}

\item{in_col}{factor in color that fill the barplot or the vertex of the network.
It can be germplasm, location or year}

\item{labels_on}{for plot_type = "network", labels to display on network for each vertex or location for map}

\item{labels_size}{for plot_type = "network", size of the labels}

\item{organize_sl}{for plot_type = "network", if TRUE, organize the network for unipart seed lots format with year 
in the chronological order on the x axis, location separated on the y axis}

\item{x_axis}{for plot_type = "barplot" and unipart seed lots network, factor on the x axis
It can be germplasm, location or year}

\item{nb_parameters_per_plot_x_axis}{for plot_type = "barplot" and unipart seed lots network, number of parameter on the x_axis}

\item{nb_parameters_per_plot_in_col}{for plot_type = "barplot" and unipart seed lots network, number of paramter by color that fill}

\item{vec_variables}{for plot_type = "barplot" and unipart seed lots network, name of a relation type to display}

\item{zoom}{zoom of the map, see ?get_map for more details}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with ggplot object.
For plot_type = "network", a list with as many elements as net with the network representation in ggplot format
For plot_type = "barplot" and bipart network, it represents the number of edges per vertex for each germplasm 
and each location.
For plot_type = "barplot" and unipart network on seed lots, it represents the number of edges per vertex for each germplasm 
and each location.
}
\description{
\code{plot.data_network} returns ggplot to visualize outputs from format_data_PPBstats()
}
\details{
For network diffusion are represented by a curve.
For organize_sl, The representation is possible if the seed_lots are under the following format : 
GERMPLASM_LOCATION_YEAR_DIGIT.

The seed_lot column from data_to_pie must refer to the seed_lot of the network.
The results is a list of two elements for each variable:
\itemize{
 \item nb_received: number of seed lots that end the relation
 \item nb_given: number of seed lots that start the relation
}

S3 method.
See the book for more details : https://priviere.github.io/PPBstats_book/describe-data-network.html
}
\seealso{
\code{\link{format_data_PPBstats.data_network}}
}
\author{
Pierre Riviere
}
