% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data_organo_napping.R
\name{plot.data_organo_napping}
\alias{plot.data_organo_napping}
\title{Plot napping object from format_data_PPBstats()}
\usage{
\method{plot}{data_organo_napping}(data, data_version = NULL,
  plot_type = c("pam", "histogramm", "barplot", "boxplot", "interaction",
  "biplot", "radar", "raster", "map"), x_axis = NULL, in_col = NULL,
  vec_variables = NULL, nb_parameters_per_plot_x_axis = 5,
  nb_parameters_per_plot_in_col = 5, labels_on = NULL, labels_size = 4,
  pie_size = 0.2)
}
\arguments{
\item{data}{The data frame coming from format_data_PPBstats() with type = "data_organo_napping"}

\item{...}{other parameters are the same as plot.data_agro.}
}
\value{
See \code{\link{plot.data_agro}}
}
\description{
\code{plot.data_napping} gets ggplot to describe the data
}
\details{
See \code{\link{plot.data_agro}}
}
\author{
Pierre Riviere
}
