% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_model_1.R
\name{plot.mean_comparisons_model_1}
\alias{plot.mean_comparisons_model_1}
\title{Get ggplot from mean_comparisons_model_1}
\usage{
\method{plot}{mean_comparisons_model_1}(x, data_version = NULL,
  ggplot.type = c("interaction", "barplot", "score"),
  nb_parameters_per_plot = 8)
}
\arguments{
\item{x}{outputs from \code{\link{mean_comparisons_model_1}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_mean_comparisons_model_1} returns ggplot from \code{\link{mean_comparisons_model_1}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{mean_comparisons_model_1}}
}
}

