% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_model_2.R
\name{plot.mean_comparisons_model_2}
\alias{plot.mean_comparisons_model_2}
\title{Get ggplot from mean_comparisons_model_2}
\usage{
\method{plot}{mean_comparisons_model_2}(x, y = NULL,
  ggplot.type = c("biplot-alpha-beta", "barplot"),
  nb_parameters_per_plot = 8)
}
\arguments{
\item{x,}{y outputs from \code{\link{mean_comparisons_model_2}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_mean_comparisons_model_2} returns ggplot from \code{\link{mean_comparisons_model_2}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{mean_comparisons_model_2}}
}
}

