% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mean_comparisons_model_anova.R
\name{plot.mean_comparisons_model_anova}
\alias{plot.mean_comparisons_model_anova}
\title{Get ggplot to visualize output from \code{\link{mean_comparisons.check_model_anova}}}
\usage{
\method{plot}{mean_comparisons_model_anova}(x, nb_parameters_per_plot = 8, ...)
}
\arguments{
\item{x}{Output from \code{\link{mean_comparisons.check_model_anova}}}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with barplot.
For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} parameters per graph.
Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction. 
The error I (alpha) and the alpha correction are displayed in the title. 
\itemize{
 \item germplasm : mean comparison for germplasm
 }
}
\description{
\code{plot.mean_comparisons_model_anova} returns ggplot to visualize outputs from \code{\link{mean_comparisons.check_model_anova}}
}
\details{
S3 method.
}
\seealso{
\code{\link{mean_comparisons.check_model_anova}}
}
\author{
Pierre Riviere
}
