% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mean_comparisons_model_bh_GxE.R
\name{plot.mean_comparisons_model_bh_GxE}
\alias{plot.mean_comparisons_model_bh_GxE}
\title{Get ggplot to visualize output from \code{\link{mean_comparisons.check_model_bh_GxE}}}
\usage{
plot.mean_comparisons_model_bh_GxE(x, y = NULL,
  plot_type = c("biplot-alpha-beta", "barplot"),
  nb_parameters_per_plot = 8, ...)
}
\arguments{
\item{x}{Output from \code{\link{mean_comparisons.check_model_bh_GxE}} for a given parameter}

\item{y}{Output from \code{\link{mean_comparisons.check_model_bh_GxE}} for a given parameter}

\item{plot_type}{"biplot-alpha-beta" or "barplot"}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with ggplot object depending on plot_type.
There are as many graph as needed with \code{nb_parameters_per_plot} parameters per graph.
\itemize{
 \item barplot :
 Letters are displayed on each bar.
 Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
 The error I (alpha) and the alpha correction are displayed in the title.
 alpha = Imp means that no differences were possible to find.
 \item biplot-alpha-beta : display the biplot with alpha_i on the x axis and beta_i on the y axis if
 x and y are coming respectively from \code{\link{mean_comparisons.check_model_bh_GxE}} for alpha and beta.
 }
}
\description{
\code{plot.mean_comparisons_model_bh_GxE} returns ggplot to visualize outputs from \code{\link{mean_comparisons.check_model_bh_GxE}}
}
\details{
S3 method.
y can not be NULL for plot_type = "biplot-alpha-beta".
See example in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-2
}
\seealso{
\code{\link{mean_comparisons.check_model_bh_GxE}}
}
\author{
Pierre Riviere
}
