% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mean_comparisons_model_bh_intra_location.R
\name{plot.mean_comparisons_model_bh_intra_location}
\alias{plot.mean_comparisons_model_bh_intra_location}
\title{Get ggplot to visualize output from \code{\link{mean_comparisons.check_model_bh_intra_location}}}
\usage{
\method{plot}{mean_comparisons_model_bh_intra_location}(x,
  plot_type = c("interaction", "barplot", "score"),
  nb_parameters_per_plot = 8, ...)
}
\arguments{
\item{x}{Output from \code{\link{mean_comparisons.check_model_bh_intra_location}}}

\item{plot_type}{"interaction", "barplot" or "score"}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with ggplot object depending on plot_type.
For each plot_type, it is a list of three elements being lists with as many elements as environment.
For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} parameters per graph.
\itemize{
\item barplot :
 \itemize{
  \item data_mean_comparisons : only environments where all MCMC converge are represented.
  Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
  The error I (alpha) and the alpha correction are displayed in the title.
  alpha = Imp means that no differences were possible to find.

  \item data_env_with_no_controls : only environments where there were no controls are represented.
  \item data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.
  }

 \item interaction :
 \itemize{
  \item data_mean_comparisons : only environments where all MCMC converge are represented.
  The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction".
  alpha = Imp means that no differences were possible to find.
  \item data_env_with_no_controls : only environments where there were no controls are represented.
  \item data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.
  }

 \item score : The score is set according to which group the entry was allocated.
 An high score means that the entry was in a group with an high mean.
 A low score means that the entry was in a group with an low mean.
 In the legend, the score goes from 1 (first group) to the number of groups of significativity.
 The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction".
 alpha = Imp means that no differences were possible to find.
 }
}
\description{
\code{plot.mean_comparisons_model_bh_intra_location} returns ggplot to visualize outputs from \code{\link{mean_comparisons.check_model_bh_intra_location}}
}
\details{
S3 method.
See example in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-1
}
\seealso{
\code{\link{mean_comparisons.check_model_bh_intra_location}}
}
\author{
Pierre Riviere
}
