% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mean_comparisons_model_home_away.R
\name{plot.mean_comparisons_model_home_away}
\alias{plot.mean_comparisons_model_home_away}
\title{Get ggplot to visualize output from \code{\link{mean_comparisons.check_model_home_away}}}
\usage{
\method{plot}{mean_comparisons_model_home_away}(x, nb_parameters_per_plot = 8, ...)
}
\arguments{
\item{x}{Output from \code{\link{mean_comparisons.check_model_home_away}}}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with barplot.
For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} parameters per graph.
Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
The error I (alpha) and the alpha correction are displayed in the title.
\itemize{
 \item version:germplasm : mean comparison for home away for each germplasm and year
 \item germplasm : mean comparison for germplasm
 \item location : mean comparison for location
 \item year : mean comparison for year
 \item version : mean comparison for home and away
 }
}
\description{
\code{plot.mean_comparisons_model_home_away} returns ggplot to visualize outputs from \code{\link{mean_comparisons.check_model_home_away}}
}
\details{
S3 method.
See in the book for mo arere details \href{https://priviere.github.io/PPBstats_book/intro-agro.html#section-freq}{here}
}
\seealso{
\code{\link{mean_comparisons.check_model_home_away}}
}
\author{
Pierre Riviere and Baptiste Rouger
}
