% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot.mean_comparisons_predict_the_past_model_bh_GxE.R
\name{plot.mean_comparisons_predict_the_past_model_bh_GxE}
\alias{plot.mean_comparisons_predict_the_past_model_bh_GxE}
\title{Get ggplot to visualize output from \code{\link{mean_comparisons.predict_the_past_model_bh_GxE}}}
\usage{
\method{plot}{mean_comparisons_predict_the_past_model_bh_GxE}(x, plot_type = "interaction", nb_parameters_per_plot = 8, ...)
}
\arguments{
\item{x}{Output from \code{\link{mean_comparisons.predict_the_past_model_bh_GxE}}}

\item{plot_type}{"interaction", "barplot" or "score"}

\item{nb_parameters_per_plot}{number of parameter per plot to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
It is the same outputs than for \code{\link{plot.mean_comparisons_model_bh_intra_location}}
}
\description{
\code{plot.mean_comparisons_predict_the_past_model_bh_GxE} returns ggplot to visualize outputs from \code{\link{mean_comparisons.predict_the_past_model_bh_GxE}}
}
\details{
S3 method.
See example in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-2
}
\seealso{
\code{\link{mean_comparisons.predict_the_past_model_bh_GxE}}
}
\author{
Pierre Riviere
}
