% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.parameter_groups.R
\name{plot.parameter_groups}
\alias{plot.parameter_groups}
\title{Get ggplot to visualize output from \code{\link{parameter_groups}}}
\usage{
\method{plot}{parameter_groups}(x, ind_to_highlight = NULL, ...)
}
\arguments{
\item{x}{Output from \code{\link{parameter_groups}}}

\item{ind_to_highlight}{individual to higlight on the PCA plot}

\item{...}{further arguments passed to or from other methods}
}
\value{
It returns list of ggplot object
  \itemize{
   \item pca : a list with three elements on the PCA on the group of parameters :
    \itemize{
     \item composante_variance : variance caught by each dimension of the PCA
     \item ind : graph of individuals
     \item var : graph of variables
    }
   \item clust : output from \code{factextra::fviz_nbclust()}. See \code{?factoextra::fviz_nbclust} for more details, 
   a list of number of cluster + 1 element
  }
}
\description{
\code{plot.parameter_groups} returns ggplot to visualize outputs from \code{\link{parameter_groups}}
}
\details{
S3 method.
See examples for AMMI, GGE and hierarchical bayesian GxE model : https://priviere.github.io/PPBstats_book/family-2.html
}
\seealso{
\code{\link{parameter_groups}}
}
\author{
Pierre Riviere
}
