% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PPBstats.R
\name{plot.PPBstats}
\alias{plot.PPBstats}
\title{Get ggplot to visualize output within the PPBstats package}
\arguments{
\item{x}{Outputs from 
\itemize{
 \item \code{\link{format_data_PPBstats}}
 \item \code{\link{check_model}} 
 \item \code{\link{mean_comparisons}} 
 \item \code{\link{parameter_groups}} 
 \item \code{\link{biplot_data}} 
 \item \code{\link{cross_validation_model_bh_GxE}}
}}

\item{...}{See details: parameters are specific to different classes}
}
\value{
It returns list of ggplot object
}
\description{
\code{plot.PPBstats} returns ggplot to visualize outputs from several functions
}
\details{
For more details see:
\itemize{
 \item \code{\link{plot.biplot_GxE}}
 \item \code{\link{plot.biplot_hedonic}}
 \item \code{\link{plot.biplot_napping}}
 \item \code{\link{plot.check_model_bh_GxE}}
 \item \code{\link{plot.check_model_bh_intra_location}}
 \item \code{\link{plot.check_model_bh_variance_intra}}
 \item \code{\link{plot.check_model_hedonic}}
 \item \code{\link{plot.check_model_napping}}
 \item \code{\link{plot.check_model_spatial}}
 \item \code{\link{plot.cross_validation_model_bh_GxE}}
 \item \code{\link{plot.data_agro}}
 \item \code{\link{plot.data_network}}
 \item \code{\link{plot.data_organo_hedonic}}
 \item \code{\link{plot.data_organo_napping}}
 \item \code{\link{plot.mean_comparisons_model_bh_GxE}}
 \item \code{\link{plot.mean_comparisons_model_bh_intra_location}}
 \item \code{\link{plot.mean_comparisons_model_GxE}}
 \item \code{\link{plot.mean_comparisons_model_hedonic}}
 \item \code{\link{plot.mean_comparisons_model_spatial}}
 \item \code{\link{plot.mean_comparisons_predict_the_past_model_bh_GxE}}
 \item \code{\link{plot.parameter_groups}}
}
}
\author{
Pierre Riviere
}
