% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{plot_check_freq_anova}
\alias{plot_check_freq_anova}
\title{plot check freq anova}
\usage{
plot_check_freq_anova(x, variable)
}
\arguments{
\item{x}{output from check_model}

\item{variable}{variable}
}
\description{
plot check freq anova
}
