% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{plot_mean_comparisons_freq_anova}
\alias{plot_mean_comparisons_freq_anova}
\title{plot mean comparisons for frequentist analysis}
\usage{
plot_mean_comparisons_freq_anova(x, variable, nb_parameters_per_plot = 8)
}
\arguments{
\item{x}{output from mean comparison}

\item{variable}{variable}

\item{nb_parameters_per_plot}{nb paramter per plot}
}
\description{
plot mean comparisons for frequentist analysis
}
