% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_variation.R
\name{post_hoc_variation}
\alias{post_hoc_variation}
\title{Plot a barplot with factors' variation for each variable}
\usage{
post_hoc_variation(list_out_check_model)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{\link{check_model}}}
}
\description{
\code{post_hoc_variation} plots a barplot with factors' variation for each variable
}
\seealso{
\itemize{
 \item \code{\link{check_model}}
 \item \code{\link{parameter_groups_model_GxE}}
}
}
