% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_variation_model_anova.R
\name{post_hoc_variation_model_anova}
\alias{post_hoc_variation_model_anova}
\title{Get dataframe for factors' variation for each variable from a list whose elements are output from \code{\link{check_model.fit_model_anova}}}
\usage{
post_hoc_variation_model_anova(list_out_check_model)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{\link{check_model}}}
}
\value{
A dataframe with the following columns: variable, factor and percentage sum of square coming from the model
}
\description{
\code{post_hoc_variation_model_anova} returns a dataframe for factors' variation for each variable from a list whose elements are output from \code{\link{check_model.fit_model_anova}}
}
\seealso{
\code{\link{post_hoc_variation}}
}
\author{
Pierre Riviere
}
