% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_variation_model_GxE.R
\name{post_hoc_variation_model_GxE}
\alias{post_hoc_variation_model_GxE}
\title{Get dataframe for factors' variation for each variable from a list whose elements are output from \code{\link{check_model_GxE}}}
\usage{
post_hoc_variation_model_GxE(list_out_check_model)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{\link{check_model}}}
}
\value{
A dataframe with the following columns: variable, factor and percentage sum of square coming from the model
}
\description{
\code{post_hoc_variation_model_GxE} returns a dataframe for factors' variation for each variable from a list whose elements are output from \code{\link{check_model_GxE}}
}
\seealso{
\code{\link{post_hoc_variation}}
}
\author{
Pierre Riviere
}
