% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{PPBdata}
\alias{PPBdata}
\title{Simulated values of thousand kernel weight (tkw) of bread wheat in a PPB programme}
\format{A data frame with 1574 rows and 11 columns}
\usage{
data(PPBdata)
}
\description{
A dataset containing one variable (tkw) and values for the following factors: germplasm, location, year, and coordinates in the trial (block, X and Y).
The different values of the parameters to create the dataset are also in columns: mu_ij, beta_jk, epsilon_ijk and sigma_j
}
\details{
tkw is in grams
}
\author{
Pierre Rivière
}
\keyword{datasets}

