% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{PPBdata2}
\alias{PPBdata2}
\title{Simulated values for three variables}
\format{A data frame with 2430 rows and 12 columns}
\usage{
data(PPBdata2)
}
\description{
A dataset containing three variables (y1, y2 and y3) and values for the following factors: germplasm, location, year and coordinates in the trial (block, X and Y).
The different values of the parameters to create y1 are also in columns: alpha_i-1, beta_i-1, and theta_j-1.
}
\author{
Pierre Rivière
}
\keyword{datasets}

