% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{package}
\name{PPBstats}
\alias{PPBstats}
\alias{PPBstats-package}
\title{PPBstats}
\description{
\code{PPBstats} is an R package to perform analysis found within PPB programmes
}
\details{
A website dedicated to PPBstats and a exhaustive tutorial 
to collaborate and use the package can be found here : 
https://priviere.github.io/PPBstats_web_site

To cite PPBstats, type citation(\"PPBstats\")
}
\author{
Pierre Riviere, Gaelle Van Franck, Olivier David and Facundo Muñoz
}
