% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats_interface.R
\name{PPBstats_interface}
\alias{PPBstats_interface}
\title{Run interface to use the package}
\usage{
PPBstats_interface(interface)
}
\arguments{
\item{interface}{Type of interface : "network", "agro", "organo" or "mol"}
}
\value{
The function returns a windows with the interface
}
\description{
\code{PPBstats_interface} runs an interface ti use the package
}
\details{
The code was taken from Dean Attali's example
https://www.r-bloggers.com/supplementing-your-r-package-with-a-shiny-app-2/
}
\author{
Pierre Riviere
}
