% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.the.past.R
\name{predict.the.past}
\alias{predict.the.past}
\title{Estimate value of a germplasm in an environment based on the FWH model.}
\usage{
\method{predict}{the.past}(out.analyse.outputs, output.format = "summary")
}
\arguments{
\item{out.analyse.outputs}{object from \code{analyse.outputs}}

\item{output.format}{Format of the outptu: "summary" or "raw". See details.}
}
\value{
The function returns a list with for each environment, the estimated value of the germplasms that were not grown in this environment.
}
\description{
\code{predict.the.past} estimates value of a germplasm in an environment based on the FWH model.
}
\details{
The estimation of the values are based ont the MCMC outputs.
More informations can be found in the vignette (type vignette("PPBstats)).

Due to memory issues, it may be better to choose output.format = "summary".
This allows caterpillar plots, barplot and interaction plots but no mean comparisons that are base on the whole MCMC.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{FWH}}, \code{\link{analyse.outputs}}
}

