% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_the_past_model_2.R
\name{predict_the_past_model_2}
\alias{predict_the_past_model_2}
\title{Estimate value of a germplasm in an environment based on the FWH model.}
\usage{
predict_the_past_model_2(out_check_model_model_2, env = NULL)
}
\arguments{
\item{out.analyse.outputs}{object from \code{analyse.outputs}}

\item{output.format}{Format of the output: "summary" or "raw". See details.}
}
\value{
The function returns a list with for each environment, the estimated value of the germplasms that were not grown in this environment.
}
\description{
\code{predict.the.past} estimates value of a germplasm in an environment based on the FWH model.
}
\details{
The estimations of the values are based on the MCMC outputs.
More informations can be found in the vignette (type vignette("PPBstats")).

Due to memory issues, it may be better to choose output.format = "summary".
This allows caterpillar plots, barplots and interaction plots but no mean comparisons that are based on the whole MCMC.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{FWH}}, \code{\link{analyse.outputs}}
}

