% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_the_past_model_bh_GxE.R
\name{predict_the_past_model_bh_GxE}
\alias{predict_the_past_model_bh_GxE}
\title{Predict values of germplasms in environments where they have not been grown based on Hierarchical Bayesian GxE model}
\usage{
predict_the_past_model_bh_GxE(out_check_model_model_bh_GxE, env = NULL)
}
\arguments{
\item{out_check_model_model_bh_GxE}{object from \code{\link{check_model.fit_model_bh_GxE}}}

\item{env}{name of the environment where the germplasm effect are predicted}
}
\value{
The function returns a MCMC for the given environment. 
This MCMC output can be used in the same way as the output from \code{\link{check_model.fit_model_bh_intra_location}}.
}
\description{
\code{predict_the_past_model_bh_GxE} predicts values of germplasms in environments where 
they have not been grown based on Hierarchical Bayesian GxE model.
}
\details{
The estimations of the values are based on the MCMC outputs.

More informations can be found in the book : https://priviere.github.io/PPBstats_book/family-2.html#model-2

It is like mu_ij effect that are estimated (as for Hierarchical Bayesian intra-location model), i.e. the effect of a germplasm in an environment.

Due to memory issues, it is better to run the function for only one environment instead of all by default.
This allows the same ggplot as for Hierarchical Bayesian intra-location model.
}
\seealso{
\itemize{
\item \code{\link{check_model}}, 
\item \code{\link{check_model.fit_model_bh_GxE}}, 
\item \code{\link{mean_comparisons}}, 
\item \code{\link{mean_comparisons.predict_the_past_model_bh_GxE}}
}
}
\author{
Pierre Riviere
}
