% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{reshape_data_split_x_axis_in_col}
\alias{reshape_data_split_x_axis_in_col}
\title{Reshape data in a list based on nb_parameters_per_plot arguments}
\usage{
reshape_data_split_x_axis_in_col(
  d,
  vec_variables,
  labels_on,
  x_axis,
  nb_parameters_per_plot_x_axis,
  in_col,
  nb_parameters_per_plot_in_col
)
}
\arguments{
\item{d}{data frame}

\item{vec_variables}{vectors of variables}

\item{labels_on}{which label to display}

\item{x_axis}{x axis}

\item{nb_parameters_per_plot_x_axis}{nb of paramters for x axis on the lpot}

\item{in_col}{in color}

\item{nb_parameters_per_plot_in_col}{nb of paramters for color on the plot}
}
\description{
Reshape data in a list based on nb_parameters_per_plot arguments
}
\details{
used in describe_data.data_agro.R and plot.data_network.R
}
