% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{split_data_for_ggplot}
\alias{split_data_for_ggplot}
\title{split data into several list for a given factor and a number of rows}
\usage{
split_data_for_ggplot(data, factor, nb_param)
}
\arguments{
\item{data}{data frame}

\item{factor}{factor on which the split is done}

\item{nb_param}{number of rows}
}
\description{
split data into several list for a given factor and a number of rows
}
