% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_data.R
\name{biplot_data}
\alias{biplot_data}
\title{Format data for biplot representation}
\usage{
biplot_data(x)
}
\arguments{
\item{x}{outputs from \code{\link{model_GxE}}, \code{\link{model_hedonic}},  \code{\link{model_napping}}}
}
\description{
\code{biplot_data} formats data for biplot representation
}
\details{
S3 method.
See for more details :
\itemize{
 \item biplot_data.check_model_GxE
 \item biplot_data.check_model_hedonic
 \item biplot_data.check_model_napping
}
}
