% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biplot_GxE.R
\name{biplot_GxE}
\alias{biplot_GxE}
\title{Compute ecovalence and format PCA results from GxE}
\usage{
biplot_GxE(out_check_model_GxE)
}
\arguments{
\item{out_check_model_GxE}{Output from \code{GxE}}
}
\value{
The function returns a list with:
\itemize{
\item variable : the variable studied
\item data_ecovalence : the ecovalence matrix
\item pca : the pca object
}
}
\description{
\code{biplot_GxE} computes ecovalence and format PCA results from \code{GxE}
}
\details{
The ecovalence is the interaction matrix squared.
}
\examples{
See the vignette for examples
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{check_model}}, \code{\link{check_model_GxE}}, \code{\link{get.ggplot}}, \code{\link{ggplot_biplot_GxE}}
}

