% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check if the model went well}
\usage{
check_model(x)
}
\arguments{
\item{x}{outputs from 
\itemize{
 \item \code{\link{model_GxE}}
 \item \code{\link{model_bh_intra_location}}
 \item \code{\link{model_bh_GxE}}
 \item \code{\link{model_spatial}}
 \item \code{\link{model_bh_variance_intra}}
 \item \code{\link{model_hedonic}}
 \item \code{\link{model_napping}}
}}
}
\description{
\code{check_model} computes tests to assess if the model went well. 
It is important to run this step before going ahead with the analysis otherwise you may make mistakes in the interpretation of the results.
}
\details{
S3 method.
See for more details :
\itemize{
 \item \code{\link{check_model.fit_model_GxE}} regarding \code{\link{model_GxE}}
 \item \code{\link{check_model.fit_model_bh_intra_location}} regarding \code{\link{model_bh_intra_location}}
 \item \code{\link{check_model.fit_model_bh_GxE}} regarding \code{\link{model_bh_GxE}}
 \item \code{\link{check_model.fit_model_spatial}} regarding \code{\link{model_spatial}}
 \item \code{\link{check_model.fit_model_bh_variance_intra}} regarding \code{\link{model_bh_variance_intra}}
 \item \code{\link{check_model.fit_model_hedonic}} regarding \code{\link{model_hedonic}}
 \item \code{\link{check_model.fit_model_napping}} regarding \code{\link{model_napping}}
}
}
\seealso{
\itemize{
 \item \code{\link{model_GxE}}
 \item \code{\link{model_bh_intra_location}}
 \item \code{\link{model_bh_GxE}}
 \item \code{\link{model_spatial}}
 \item \code{\link{model_bh_variance_intra}}
 \item \code{\link{model_hedonic}}
 \item \code{\link{model_napping}}
 \item \code{\link{mean_comparisons}}
 \item \code{\link{biplot_data}}
 \item \code{\link{plot.PPBstats}}
}
}
