% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model_GxE.R
\name{check_model_GxE}
\alias{check_model_GxE}
\title{Check if the GxE model went well}
\usage{
check_model_GxE(out_GxE)
}
\arguments{
\item{out_GxE}{outputs from GxE function}
}
\description{
\code{check_model_GxE} compute test to assess if the GxE model went well
}
\details{
See check_model for further information
}
\seealso{
\code{\link{GxE}}, \code{\link{check_model}}
}

