% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model_model_1.R
\name{check_model_model_1}
\alias{check_model_model_1}
\title{Check if the model_1 model went well}
\usage{
check_model_model_1(out_model_1)
}
\arguments{
\item{out_model_1}{outputs from model_1 function}
}
\value{
See check_model
}
\description{
\code{check_model_model_1} compute test to assess if the model_1 model went well
}
\details{
See check_model
}
\seealso{
\code{\link{model_1}}, \code{\link{check_model}}
}

