% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_model_model_2.R
\name{check_model_model_2}
\alias{check_model_model_2}
\title{Check if the model_2 model went well}
\usage{
check_model_model_2(out_model_2)
}
\arguments{
\item{out_model_2}{outputs from \code{\link{model_2}}}
}
\value{
See \code{\link{check_model}}
}
\description{
\code{check_model_model_2} computes tests to assess if the model_2 model went well
}
\details{
See \code{\link{check_model}}
}
\seealso{
\itemize{
\item \code{\link{model_2}},
\item \code{\link{check_model}}
}
}

