% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{common_functions}
\alias{common_functions}
\alias{check_data_vec_variables}
\title{Some functions used in several functions of PPBstats}
\usage{
check_data_vec_variables(data, vec_variables)
}
\description{
This file group several functions used in several functions of PPBstats
}
\author{
Pierre Riviere
}
