% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{common_functions}
\alias{common_functions}
\alias{splitter_d}
\title{Some functions used in one or several functions of PPBstats}
\usage{
splitter_d(data, .variables = NULL, drop = TRUE)
}
\arguments{
\item{data}{data}

\item{.variables}{.variables}

\item{drop}{drop}
}
\description{
This file group functions used in several functions of PPBstats
}
\author{
Pierre Riviere

copy paste from plyr:::splitter_d in order to avoid NOTE in the package check ... I know it is quite dirty to do that !
}
