% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross_validation_model_2.R
\name{cross_validation_model_2}
\alias{cross_validation_model_2}
\title{Run complete cross validation with model 2 (FWH)}
\usage{
cross_validation_model_2(data, variable, nb_iterations = 1e+05,
  mc.cores = 1)
}
\arguments{
\item{data}{The data frame on which the model will be run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{The variable on which runs the model}

\item{nb_iterations}{Number of iterations of the MCMC}

\item{mc.cores}{The number of cores used for parallelisation of the computing}
}
\value{
The function returns a list with
\itemize{
 \item A data frame with real values and estimated values by cross validation
 \item A list with the plot of the regression and the anova model
 \item The percentage of confidence
 }
}
\description{
\code{cross.validation.FWH} runs complete cross validation with model 2 (FWH)
}
\details{
The convergence is not checked for each validation. If the parameters in the FWH converge (cf \code{analyse.outputs}), then it is assumed that the FWH in the cross validation converge as well.

The model is run on data sets where germplasms are on at least three environments so the smallest data set where the cross validation is run has germplasms on at least two environments.

Parallelization is done with the function \code{parallel::mclapply}.

The percentage of confidence is estimated from the bias of estimated values in relation to real values.
A t.test is performed with the null hypothesis H0: the bias = 0.
More informations can be found in the vignette (type vignette("PPBstats")).
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{FWH}}, \code{\link{predict.the.past}}
}

