% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_GxE}
\alias{data_GxE}
\title{Simulated data set for three variables to test GxE model}
\format{A data frame with 180 rows and 14 columns}
\usage{
data(data_GxE)
}
\description{
A dataset containing five variables (y1, y2, y3, desease, vigor) 
with their respective dates for the first three (y1$date, y2$date and y3$date)
and values for the following factors: location, year, germplasm, block, X and Y.
}
\author{
Pierre Rivière
}
\keyword{datasets}
