% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{data}
\name{data_model_spatial}
\alias{data_model_spatial}
\title{Data set for one variable to test spatial model}
\format{A data frame with 211 rows and 10 columns}
\usage{
data(data_model_spatial)
}
\description{
A dataset containing three variables y1, y2 and y3
the following factors: seed_lot, location, year, germplasm, block, X and Y.
}
\author{
Pierre Rivière
}
\keyword{datasets}
