% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe the data set in order to choose the appropriate analysis to carry out}
\usage{
describe_data(data, vec_variables, nb_parameter_per_grid = 5)
}
\arguments{
\item{data}{The data frame. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{vec_variables}{Vector of variables to describe}

\item{nb_parameter_per_grid}{Nomber of parameter on each histogram on the gird}
}
\value{
The function returns a list with,
\itemize{
 \item A summary of the whole data set
 \item for each variable, a list with :
  \itemize{
  \item A presence.abscence plot
  \item A list with histogram for
   \itemize{
    \item germplasm
    \item location
    \item year
    }
  \item A list with boxplot, containg a list with plot and outliers, for
   \itemize{
    \item germplasm
    \item location
    \item year
    }
  }
}
}
\description{
\code{describe_data} describes the data set in order to choose the appropriate analysis to carry out
}
\author{
Pierre Riviere
}

