% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.data_organo_hedonic.R
\name{format_data_PPBstats.data_organo_hedonic}
\alias{format_data_PPBstats.data_organo_hedonic}
\title{Check and format the data to be used by PPBstats functions for hedonic analyses}
\usage{
format_data_PPBstats.data_organo_hedonic(data, threshold)
}
\arguments{
\item{data}{The data frame with the following columns : sample, juges, note, descriptors, germplasm, location. 
The descriptors must be separated by ";". Any other column can be added as supplementary variables.}

\item{threshold}{number of occurence of descriptors <= threshold are kept}
}
\description{
\code{format_data_PPBstats.data_organo_hedonic} checks and formats the data to be used by PPBstats functions for hedonic analyses
}
\details{
See the book for more details \href{https://priviere.github.io/PPBstats_book/hedonic.html#format-the-data-8}{here}.
}
\seealso{
\code{\link{format_data_PPBstats}}
}
\author{
Pierre Riviere
}
