% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.R
\name{format_data_PPBstats}
\alias{format_data_PPBstats}
\title{Check and format the data to be used in PPBstats functions}
\usage{
format_data_PPBstats(data, type = "data_agro")
}
\arguments{
\item{data}{The data frame. See details.}

\item{type}{type of format : "data_network", "data_agro" or "data_agro_version". See details.}
}
\value{
The function returns the data with the right format (i.e. class)
}
\description{
\code{format_data_PPBstats} checks and formats the data to be used in PPBstats functions
}
\details{
\itemize{
 \item For type = "data_network" :
 \item For type = "data_agro" : 
 It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.
 The variables can be linked to their corresponding dates. 
 The dates are associated to their corresponding variable by $.
 For example the date associated to variable y1 is y1$date.
 \item For type = "data_agro_version" : 
 It should have the following columns: c("year", "germplasm", "location", "group", "version").
 The group refers to an id that contains two different versions.
 For example for group 1, there is version 1 and 2. 
 See data(data_version) for an example.
}
}
\author{
Pierre Riviere
}
