% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_PPBstats.R
\name{format_data_PPBstats}
\alias{format_data_PPBstats}
\title{Check and format the data to be used by PPBstats functions}
\usage{
format_data_PPBstats(data, type, threshold, network_part = c("unipart",
  "bipart"), network_split = c("germplasm", "relation_year_start"),
  vertex_type = NULL)
}
\arguments{
\item{data}{The data frame to format}

\item{type}{type of format : 
\itemize{
 \item data_network
 \item data_agro
 \item data_agro_version
 \item data_agro_version_SR
 \item data_agro_version_MR
 \item data_organo_napping
 \item data_organo_hedonic
 }}

\item{threshold}{For type = data_organo, number of occurence of descriptors <= threshold are kept}

\item{network_part}{For type = "data_network", element of the network, it can be "unipart" or "bipart"}

\item{network_split}{For type = "data_network" and network_part = "unipart" and 
vertex_type = "location", split of the data that can be "germplasm" or "relation_year_start"}

\item{vertex_type}{For type = "data_network",
\itemize{
 \item for unipart network : "seed_lots" or "location"
 \item for bipart network : c("germplasm", "location")
 }}
}
\description{
\code{format_data_PPBstats} checks and formats the data to be used by PPBstats functions
}
\details{
See for more details :
\itemize{
 \item \code{\link{format_data_PPBstats.data_network}} 
 \item \code{\link{format_data_PPBstats.data_agro}} 
 \item \code{\link{format_data_PPBstats.data_agro_version}} 
 \item \code{\link{format_data_PPBstats.data_organo_napping}} 
 \item \code{\link{format_data_PPBstats.data_organo_hedonic}} 
}
}
\author{
Pierre Riviere
}
