% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{format_organo}
\alias{format_organo}
\title{format data for organoleptic analysis}
\usage{
format_organo(data, threshold, var_sup)
}
\arguments{
\item{data}{data frame}

\item{threshold}{threshold}

\item{var_sup}{supplementary variables}
}
\description{
format data for organoleptic analysis
}
\details{
see \code{\link{format_data_PPBstats.data_organo_napping}} and 
\code{\link{format_data_PPBstats.data_organo_hedonic}}
}
