% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.significant.groups.R
\name{get.significant.groups}
\alias{get.significant.groups}
\title{Get significant groups of means that are significantly different from each others
 for a set of parameters based on MCMC outputs}
\usage{
get.significant.groups(Mpvalue, MCMC, alpha = 0.05, p.adj = NULL)
}
\arguments{
\item{Mpvalue}{Square matrix with the probability of having a common distribution for each pair of parameter. It comes from \code{\link{comp.parameters}}}

\item{MCMC}{MCMC. It is a data frame.}

\item{alpha}{The level of type one error. 0.05 (5\%) by default}

\item{p.adj}{NULL For no adjustement of the pvalue. "soft.bonf" for a soft bonferonni correction to take into account multiple comparisons (alpha / nb of parameters).}
}
\value{
The function returns a data frame of four columns and (nb of parameters) rows. Columns are parameters, mean, median and groups
}
\description{
\code{get.significant.groups} gets significant groups of means that are significantly different 
from each others for a set of parameters based on MCMC outputs
}
\seealso{
\itemize{
\item \code{\link{mean_comparisons_model_1}}
\item \code{\link{mean_comparisons_model_2}}
}
}
\author{
Pierre Riviere
}
