% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ggplot.R
\name{get_ggplot}
\alias{get_ggplot}
\title{Get ggplot objects to visualize output from several functions}
\usage{
get_ggplot(data, data_2 = NULL, data_version = NULL,
  ggplot.type = "interaction", nb_parameters_per_plot = 8)
}
\arguments{
\item{data}{Outputs from
\itemize{
 \item \code{check_model} or
 \item \code{mean_comparisons} or
 \item \code{parameter_groups} or
 \item \code{biplot_GxE} or
 \item \code{cross_validation_model_2}
}}

\item{data_2}{Outputs from \code{mean_comparisons} from \code{check_model} from \code{model_2}.
This argument is used only when argument data comes from \code{mean_comparisons} from \code{check_model} from \code{model_2}.}

\item{data_version}{data set with the following columns: "year", "germplasm", "location", "group", "version". The group refers to an id that contains two different versions. For example for group 1, there is version 1 and 2. See data(data_version) for an example.
This argument is used only for \code{mean_comparisons} from
\itemize{
 \item \code{check_model} from \code{model_1}
 \item \code{check_model} from \code{predict_the_past_model_2}
}}

\item{ggplot.type}{The type of plot you wish.
\itemize{
 \item "barplot" for mean_comparisons from
 \itemize{
  \item check_model from GxE, model_1, model_2
  \item predict_the_past_model_2
  }
 \item "interaction.plot" for mean_comparisons from
 \itemize{
  \item check_model from model_1
  \item predict_the_past_model_2
  }
 \item "score" for mean_comparisons from
 \itemize{
  \item check_model from model_1
  \item predict_the_past_model_2
  }
 \item "biplot-alpha-beta" for mean_comparisons from check_model from model_2
 }

There are automatic plots generated for several functions:
\itemize{
 \item \code{biplot_GxE}
 \item \code{check_model}
 \item \code{parameter_groups}
 \item \code{cross_validation_model_2}
 }
See details for more information.}

\item{nb_parameters_per_plot}{The number of parameters per plot to facilitate the visualization}
}
\value{
\itemize{


 \item check_model
  \itemize{

   \item from GxE :
    \itemize{
     \item residuals
      \itemize{
       \item histogramm : histogramm with the distribution of the residuals
       \item qqplot
      }
     \item variability_repartition : pie with repartition of SumSq for each factor
     \item variance_intra_germplasm : repartition of the residuals for each germplasm (see Details for more information)
     \item pca_composante_variance : variance caught by each dimension of the PCA run on the interaction matrix
    }

   \item from model_1 :
    \itemize{
     \item sigma_j_gamma : mean of each sigma_j displayed on the Inverse Gamma distribution. The first graph represent all the sigma_j, the other graph represent \code{nb_parameters_per_plot} sigma_j per graph.
     \item mu_ij : distribution of each mu_ij in a list with as many elements as environment. For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} mu_ij per graph.
     \item beta_jk : distribution of each beta_jk in a list with as many elements as environment. For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} beta_jk per graph.
     \item sigma_j : distribution of each sigma_j. There are as many graph as needed with \code{nb_parameters_per_plot} sigma_j per graph.
     \item epsilon_ijk : standardised residuals distribution.
     If the model went well it should be between -2 and 2.
     \item mcmc_not_converge_traceplot_density : a list with the plots of trace and density to check the convergence of the two MCMC only for chains that are not converging thanks to the Gelman-Rubin test.
     If all the chains converge, it is NULL
    }

   \item from model_2 :
    \itemize{
     \item alpha_i : distribution of each alpha_i. There are as many graph as needed with \code{nb_parameters_per_plot} alpha_i per graph.
     \item beta_i : distribution of each beta_i. There are as many graph as needed with \code{nb_parameters_per_plot} beta_i per graph.
     \item theta_j : distribution of each theta_j. There are as many graph as needed with \code{nb_parameters_per_plot} theta_j per graph.
     \item epsilon_ij : standardised residuals distribution.
     If the model went well it should be between -2 and 2.
     \item mcmc_not_converge_traceplot_density : a list with the plots of trace and density to check the convergence of the two MCMC only for chains that are not converging thanks to the Gelman-Rubin test.
     If all the chains converge, it is NULL
    }

  }


 \item mean_comparisons
  \itemize{
   \item from GxE, a list with barplot.
   For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} perameters per graph.
   Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
   The error I (alpha) and the alpha correction are displayed in the title.
    \itemize{
     \item germplasm : mean comparison for germplasm
     \item location : mean comparison for location
     \item year : mean comparison for year
    }

   \item from model_1, a list with ggplot object depending on ggplot.type.
   For each ggplot.type, it is a list of three elements being lists with as many elements as environment.
       For each element of the list, there are as many graph as needed with \code{nb_parameters_per_plot} perameters per graph.
    \itemize{
     \item barplot :
      \itemize{
       \item data_mean_comparisons : only environments where all MCMC converge are represented.
       Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
       The error I (alpha) and the alpha correction are displayed in the title.
       alpha = Imp means that no differences were possible to find.
       \item data_env_with_no_controls : only environments where there were no controls are represented.
       \item data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.
      }
      Note that when using data_version, the pvalue is computed based on the MCMC.
      For data that did not converge or without environments, it is a \code{t.test} which is perform.
     \item interaction :
      \itemize{
       \item data_mean_comparisons : only environments where all MCMC converge are represented.
       The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction".
       alpha = Imp means that no differences were possible to find.
       \item data_env_with_no_controls : only environments where there were no controls are represented.
       \item data_env_whose_param_did_not_converge : only environments where MCMC did not converge are represented.
      }
     \item score : The score is set according to which group the entry was allocated.
       An high score means that the entry was in a group with an high mean.
       A low score means that the entry was in a group with an low mean.
       The error I (alpha) and the alpha correction are displayed in the x.axis under the form "alpha | alpha correction".
       alpha = Imp means that no differences were possible to find.
    }


   \item from model_2, a list with ggplot object depending on ggplot.type.
   There are as many graph as needed with \code{nb_parameters_per_plot} parameters per graph.
    \itemize{
     \item barplot :
       Letters are displayed on each bar. Parameters that do not share the same letters are different regarding type I error (alpha) and alpha correction.
       The error I (alpha) and the alpha correction are displayed in the title.
       alpha = Imp means that no differences were possible to find.
     \item biplot-alpha-beta : display the biplot with alpha_i on the x axis and beta_i on the y axis.
    }


   \item predict_the_past_model_2, a list with ggplot object depending on ggplot.type.
   It is the same outputs than for model_1

  }


 \item parameter_groups
  \itemize{
   \item pca : a list with three elements on the PCA on the group of parameters :
    \itemize{
     \item composante_variance : variance caught by each dimension of the PCA
     \item ind : graph of individuals
     \item var : graph of variables
    }
   \item clust
    \itemize{
     \item nb_k : output from factextra::fviz_nbclust(data, kmeans, method = "silhouette")
     See ?factoextra::fviz_nbclust for more details
     \item pca : output from factextra::fviz_cluster(km.res, data, frame.type = "norm")
     See ?factoextra::fviz_cluster for more details
    }
  }


 \item biplot_GxE
  \itemize{
   \item ecovalence :
   \item biplot, a list of three elements :
    \itemize{
     \item which_won_where : plot to assess which germplasm win in which location (see Details for more information).
     \item mean_vs_stability : a list of two elements :
      \itemize{
       \item mean :assess mean of each germplasm (see Details for more information).
       \item stability : assess stability of each germplasm (see Details for more information).
      }
     \item discrimitiveness_vs_representativeness : a list of two elements :
      \itemize{
       \item discrimitiveness : assess discrimitiveness of each location (see Details for more information).
       \item representativeness : assess representativeness of each location (see Details for more information).
      }
    }
  }


 \item cross_validation_model_2
  \itemize{
   \item plot : plot estimated.value = f(observed.value).
   The probability mean = 0 is display (see Details for more information).
   \item regression : output of the model observed.value = a x estimated.value + b
  }

}
}
\description{
\code{get_ggplot} returns ggplot objects to visualize outputs from several functions
}
\details{
Below are some details on the construction of the plots :

\itemize{
 \item check_model from GxE :
  \itemize{
   \item variance_intra_germplasm represents the repartition of the residuals for each germplasm.
   This has to been seen with caution:
   \itemize{
    \item If germplasm have no intra-germplasm variance (i.e. pure line or hybrides) then the distribution of each germplasm represent only the micro-environmental variation.
    \item If germplasm have intra-germplasm variance (i.e. population such as landraces for example) then the distribution of each germplasm represent the micro-environmental variation plus the intra-germplasm variance.
    With the hypothesis than the micro-environmental variation is equaly distributed on all the individuals (i.e. all the plants), the distribution of each germplasm represent the intra-germplasm variance.
   }
 }

 \item check_model from model_1 and model_2
 \itemize{
  \item mcmc_not_converge_traceplot_density : If you wish exhaustive information, look at \code{ggmcmc::ggmcmc} with \code{ggmcmc(out_model$MCMC)}.
  But be careful with the size of your MCMC output which are often too big to be performed in R.
 }

 \item biplot_GxE
  \itemize{
   \item which_won_where : plot to assess which germplasm win in which location (see Details for more information).
   \item mean_vs_stability : a list of two elements :
    \itemize{
     \item mean :assess mean of each germplasm (see Details for more information).
     \item stability : assess stability of each germplasm (see Details for more information).
    }
   \item discrimitiveness_vs_representativeness : a list of two elements :
    \itemize{
     \item discrimitiveness : assess discrimitiveness of each location (see Details for more information).
     \item representativeness : assess representativeness of each location (see Details for more information).
    }
   }

 \item cross_validation_model_2
  \itemize{
   \item plot : plot estimated.value = f(observed.value).
   The probability mean = 0 is coming from a t.test performed with the null hypothesis H0: the bias of estimated values in relation to real values = 0.
  }

}
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
 \item \code{check_model}
 \item \code{mean_comparisons}
 \item \code{parameter_groups}
 \item \code{biplot_GxE}
 \item \code{cross_validation_model_2}
}
}

