% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_biplot_GxE.R
\name{ggplot_biplot_GxE}
\alias{ggplot_biplot_GxE}
\title{Get ggplot from biplot_GxE}
\usage{
ggplot_biplot_GxE(out_biplot_GxE)
}
\arguments{
\item{out_biplot_GxE}{outputs from \code{\link{biplot_GxE}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_biplot_GxE} returns ggplot from \code{\link{biplot_GxE}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{biplot_GxE}}
}
}

