% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_check_model_model_2.R
\name{ggplot_check_model_model_2}
\alias{ggplot_check_model_model_2}
\title{Get ggplot objects from check_model_model_2}
\usage{
ggplot_check_model_model_2(out_check_model_model_2,
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_check_model_model_2}{outputs from check_model_model_2 function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_check_model_model_2} returns ggplot objects from check_model_model_2
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{check_model_model_2}}
}

