% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_discrimitiveness_vs_representativeness.R
\name{ggplot_discrimitiveness_vs_representativeness}
\alias{ggplot_discrimitiveness_vs_representativeness}
\title{Return a "discrimitiveness vs representativeness" ggplot based on PCA object}
\usage{
ggplot_discrimitiveness_vs_representativeness(res.pca)
}
\arguments{
\item{res.pca}{output from \code{\link{biplot_data.check_model_GxE}}}
}
\value{
A ggplot object.
}
\description{
\code{ggplot_discrimitiveness_vs_representativeness} returns a "discrimitiveness vs representativeness" ggplot based on PCA object
}
\details{
It is a list of three elements.
The two first elements where the plots are drawn under the same construction.
A red circle indicates the position of the average location, which is defined by the average Dim1 and Dim2 scores across all locations.
An arrow points this average 'virtual' location.
A red line passes through the biplot origin and the average location (known as the Average-Tester Axis (ATA)).

\itemize{
 \item discrimitiveness : assess discrimitiveness of each location.
 The score represents the discrimination : the higher the vector, the more discriminating the location.

 \item representativeness : assess representativeness of each location.
 The score represents the representativeness : the higher the score, the less representative the location.

 \item discrimitiveness_vs_representativeness :  represent discrimitiveness versus representativeness.
 The location combining better score (i. e.discrimination and representativeness) are the ones that could be used to test germplasms as they are more representative of all the locations.
 This has to be done severals year to get robust results.
 The highest the score, the more disciminative and representative the location.
 }
}
\seealso{
\code{\link{plot.PPBstats}}
\code{\link{plot.biplot_GxE}}
}
\author{
Pierre Riviere
}
