% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_GxE.R
\name{ggplot_mean_comparisons_GxE}
\alias{ggplot_mean_comparisons_GxE}
\title{Get ggplot objects from mean_comparisons_GxE}
\usage{
ggplot_mean_comparisons_GxE(mean_comparisons_GxE, ggplot.type = "barplot",
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_mean_comparisons_GxE}{outputs from mean_comparisons_GxE function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_mean_comparisons_GxE} returns ggplot objects from mean_comparisons_GxE
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{mean_comparisons_GxE}}
}

