% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_predict_the_past_model_2.R
\name{ggplot_mean_comparisons_predict_the_past_model_2}
\alias{ggplot_mean_comparisons_predict_the_past_model_2}
\title{Get ggplot objects from mean_comparisons_predict_the_past_model_2}
\usage{

  ggplot_mean_comparisons_predict_the_past_model_2(out_mean_comparisons_predict_the_past_model_2,
  data_version = NULL, ggplot.type = "interaction",
  nb_parameters_per_plot = 10)
}
\arguments{
\item{out_mean_comparisons_predict_the_past_model_2}{outputs from mean_comparisons_predict_the_past_model_2 function}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_mean_comparisons_predict_the_past_model_2} returns ggplot objects from mean_comparisons_predict_the_past_model_2
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{mean_comparisons_predict_the_past_model_2}}
}

