% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_vs_stability.R
\name{ggplot_mean_vs_stability}
\alias{ggplot_mean_vs_stability}
\title{Get "mean vs stability" ggplot objects from PCA on interaction matrix}
\usage{
ggplot_mean_vs_stability(res.pca)
}
\arguments{
\item{res.pca}{PCA object on interaction matrix}
}
\value{
See get_ggplot
}
\description{
\code{ggplot_mean_vs_stability} returns ggplot objects from PCA on interaction matrix
}
\details{
See get_ggplot
}
\seealso{
\code{\link{get_ggplot}}, \code{\link{biplot_GxE}}
}

