% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_parameter_groups.R
\name{ggplot_parameter_groups}
\alias{ggplot_parameter_groups}
\title{Get ggplot objects from parameter_groups}
\usage{
ggplot_parameter_groups(out_parameter_groups)
}
\arguments{
\item{out_parameter_groups}{outputs from \code{\link{parameter_groups}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_parameter_groups} returns ggplot objects from \code{\link{parameter_groups}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{parameter_groups}}
}
}

