% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_which_won_where.R
\name{ggplot_which_won_where}
\alias{ggplot_which_won_where}
\title{Get "which won where" ggplot from PCA on interaction matrix}
\usage{
ggplot_which_won_where(res.pca)
}
\arguments{
\item{res.pca}{PCA object on interaction matrix}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_which_won_where} returns ggplot from PCA on interaction matrix
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}}, 
\item 
}\code{\link{biplot_GxE}}
}
