% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_which_won_where.R
\name{ggplot_which_won_where}
\alias{ggplot_which_won_where}
\title{Return a "which won where" ggplot based on PCA object}
\usage{
ggplot_which_won_where(res.pca)
}
\arguments{
\item{res.pca}{output from \code{\link{biplot_data.check_model_GxE}}}
}
\value{
A ggplot object.
}
\description{
\code{ggplot_which_won_where} returns a "which won where" ggplot based on PCA object
}
\details{
The plot to assess which germplasm win in which location.
A polygon is drawn connecting germplasm which are located furthest away from the biplot origin such that all other entries are contained within the polygon;
Perpendicular lines are drawn to each side of the polygon and start from biplot origin.
A sector is the triangle area formed by two perpendicular lines.
The germplasms which have the largest value in a sector "win" in the location present in that sector.
The information is summarized in the legend of the plot.
}
\seealso{
\code{\link{plot.PPBstats}}
\code{\link{plot.biplot_GxE}}
}
\author{
Pierre Riviere
}
