% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GxE.R
\name{GxE}
\alias{GxE}
\title{Run AMMI or GGE model on a data set}
\usage{
GxE(data, variable, gxe_analysis = NULL)
}
\arguments{
\item{data}{data with at least the following columns: year, location, germplasm, block, X, Y}

\item{gxe_analysis}{the analysis to carry out: "AMMI" or "GGE"}

\item{vec_variables}{vectors of variables to analyse}
}
\value{
The function returns a list with three elements:
\itemize{
 \item variable
 \item ANOVA a list with five elements:
  \itemize{
   \item model
   \item anova_model
   \item germplasm_effects a list of two elements:
    \itemize{
     \item effects
     \item intra_variance
    }
   \item location_effects
   \item interaction_matrix
  }

 \item PCA : PCA object from FactoMineR
}
}
\description{
\code{GxE} runs AMMI or GGE model
}
\details{
scaling for interaction.matrix is not useful as the column mean is equal to 0 because of model constraints and all the values are regarding one variable, so it is possible to compare it into the PCA.
}
\author{
Pierre Riviere
}

