% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GxE.R
\name{GxE}
\alias{GxE}
\title{Run GxE model on a data set}
\usage{
GxE(data, variable, gxe_analysis = NULL)
}
\arguments{
\item{data}{data with at least the following columns: year, location, germplasm, block, X, Y}

\item{gxe_analysis}{the analysis to carry out: "AMMI" or "GGE"}

\item{vec_variables}{vectors of variables to analyse}
}
\value{
The function returns two lists with
\itemize{
 \item for each variable a list with
  \itemize{
   \item "descriptive" being a list with
   \itemize{
    \item germplasm containing a boxplot for each germplasm
    \item location containing a boxplot for each location
    \item interaction containing an interaction plot
    }

   \item "ANOVA" being a list with
    \itemize{
     \item "anova_model"
     \item "residuals" being a list with
      \itemize{
       \item "distribution"
       \item "QQplot"
      }
     \item "variability_repartition" a pie chart with variability repartition
     \item "location_effects" being a list with
     \itemize{
      \item a vector with location effects
      \item a barplot with groups based on LSD interval
     }
     \item "germplasm_effects"
     \itemize{
      \item a vector with germplasm effects
      \item a vector with intra germplasm variance estimation
      \item a barplot with groups based on LSD interval
      \item a boxplot with intra germplasm variance estimation per germplasm
     }
     \item "interaction_matrix" the interaction matrix on which is based the PCA
    }

   \item PCA being a list with
    \itemize{
     \item "ecovalence"
     \item "biplot" a list with
     \itemize{
      \item "variation_dim"
      \item "which_won_where"
      \item "mean_vs_stability" a list with two elements: "mean" and "stability"
      \item "discrimitiveness_vs_representativeness"
     }
   }
 }

 \item a post GxE analysis with
  \itemize{
   \item "barplot_variation_repartition"
   \item "PCA_G_effect"
   \item "PCA_intraG_effect"
   \item "PCA_E_effect"
  }
 }
}
\description{
\code{GxE} runs AMMI or GGE model on a data set and gives back results on germplasms, locations, years and GxE or G+GxE interactions effects
}
\details{
See the vignette for more details on the models
}
\author{
Pierre Riviere
}

