% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GxE.R
\name{GxE}
\alias{GxE}
\title{Run AMMI or GGE model}
\usage{
GxE(data, variable, gxe_analysis = NULL)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{variable}{variable to analyse}

\item{gxe_analysis}{the analysis to carry out: "AMMI" or "GGE"}
}
\value{
The function returns a list with three elements:
\itemize{
 \item info : a list with variable and gxe_analysis
 \item ANOVA a list with five elements:
  \itemize{
   \item model
   \item anova_model
   \item germplasm_effects a list of two elements:
    \itemize{
     \item effects
     \item intra_variance
    }
   \item location_effects
   \item interaction_matrix
  }

 \item PCA : PCA object from FactoMineR
}
}
\description{
\code{GxE} runs AMMI or GGE model
}
\details{
scaling for interaction.matrix is not useful as the column mean is equal to 0 because of model constraints and all the values are regarding one variable, so it is possible to compare it into the PCA.
}
\author{
Pierre Riviere
}
\seealso{
\itemize{
\item \code{\link{GxE_build_interaction_matrix}},
\item \code{\link{check_model}},
\item \code{\link{check_model_model_1}}
}
}

