% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GxE_build_interaction_matrix.R
\name{GxE_build_interaction_matrix}
\alias{GxE_build_interaction_matrix}
\title{Compute interaction matrix}
\usage{
GxE_build_interaction_matrix(data, gxe_analysis)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{gxe_analysis}{the analysis to carry out: "AMMI" or "GGE"}
}
\value{
The function returns the interaction matrix
}
\description{
\code{GxE_build_interaction_matrix} computes interaction matrix for AMMI or GGE model
}
\details{
The computation is inspired from \code{agricolae::AMMI}.
Another way to do has been explore but did not succeed regarding missing data.
The code is inside the function within a if(FALSE){}.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{GxE}}
}

