% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{HA_to_LF}
\alias{HA_to_LF}
\title{transform home away data to local foreign data}
\usage{
HA_to_LF(data_version_HA)
}
\arguments{
\item{data_version_HA}{data home away}
}
\value{
a data frame of class data_version_LF
}
\description{
transform home away data to local foreign data
}
\details{
change home to local and away to foreign
}
