% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparisons_GxE.R
\name{mean_comparisons.check_model_GxE}
\alias{mean_comparisons.check_model_GxE}
\title{Get mean comparisons from check_model_GxE}
\usage{
\method{mean_comparisons}{check_model_GxE}(x, alpha = 0.05, p.adj = "none")
}
\arguments{
\item{x}{}

\item{alpha}{}

\item{p.adj}{}
}
\value{
See \code{\link{mean_comparisons}}
}
\description{
\code{mean_comparisons_GxE} performs mean comparisons from object coming from \code{\link{check_model_GxE}}
See \code{\link{mean_comparisons}} for more information.
}
\details{
See \code{\link{mean_comparisons}}
}
\seealso{
\itemize{
 \item \code{\link{mean_comparisons}}
 \item \code{\link{get_ggplot}}
}
}
