% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_hedonic.R
\name{model_hedonic}
\alias{model_hedonic}
\title{Run hedonic analysis}
\usage{
model_hedonic(data)
}
\arguments{
\item{data}{data is a data frame. It should come from \code{\link{format_data_PPBstats.data_organo_hedonic}}}
}
\value{
It returns a list with three elements:
\itemize{
 \item model : the result of the anova run on note
 \item CA the result of the correspondance analysis run on the data set with the supplementary variables with \code{FactoMineR::CA}.
 \item HCPC the result of the correspondane analysis run on the data set with the supplementary variables with \code{FactoMineR::PCA} follow by \code{FactoMineR::HCPC}.
 It is a list of three elements:
 \itemize{
  \item res.pca the results of the PCA
  \item res.hcpc the results of the HCPC
  \item clust the cluster found with the HCPC
 }
}
}
\description{
\code{model_hedonic} runs hedonic analysis
}
\details{
An anova is run on "note".
A Correspondence Analysis is run with FactoMineR::CA

More information can be found in the book : https://priviere.github.io/PPBstats_book/hedonic.html
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_hedonic}}
}
}
