% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_home_away.R
\name{model_home_away}
\alias{model_home_away}
\title{Run Home away model}
\usage{
model_home_away(data, data_version, variable)
}
\arguments{
\item{data}{The data frame on which the model is run. It should have at least the following columns : c("year", "germplasm", "location", "block", "X", "Y", "..."), with "..." the variables.}

\item{data_version}{}

\item{variable}{variable to analyse}
}
\value{
The function returns a list with three elements :
\itemize{
 \item info : a list with variable
 \item ANOVA a list with five elements:
  \itemize{
   \item model
   \item anova_model
   \item germplasm_effects a list of two elements:
    \itemize{
     \item effects
     \item intra_variance
    }
   \item location_effects
   \item interaction_matrix
  }
 
}
}
\description{
\code{model_home_away} runs home away model
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_home_away}}
}
}
\author{
Pierre Riviere and Gaelle Van Frank
}
