% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_home_away.R
\name{model_home_away}
\alias{model_home_away}
\title{Run Home away model}
\usage{
model_home_away(data, data_version, variable)
}
\arguments{
\item{data}{The data frame on which the model is run. 
It should come from \code{\link{format_data_PPBstats.data_agro}}}

\item{data_version}{It should come from \code{\link{format_data_PPBstats.data_agro_version}} with home and away in version}

\item{variable}{variable to analyse}
}
\value{
The function returns a list with three elements :
\itemize{
 \item info : a list with variable
 \item ANOVA a list with two elements:
  \itemize{
   \item model
   \item anova_model
  }
}
}
\description{
\code{model_home_away} runs home away model
}
\details{
Find details in the book \href{https://priviere.github.io/PPBstats_book/family-4.html#family-4}{here}.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_home_away}}
}
}
\author{
Pierre Riviere and Gaelle Van Frank
}
