% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_napping.R
\name{model_napping}
\alias{model_napping}
\title{Run napping analysis}
\usage{
model_napping(data)
}
\arguments{
\item{data}{data frame coming from format_data_PPBstats().
It has the following columns: sample, juges, X, Y, descriptors, germplasm, location. 
The descriptors must be separated by ";"}
}
\value{
The MFA object
}
\description{
\code{model_napping} runs napping analysis
}
\details{
The Multiple Factor Analysis is run with FactoMineR::MFA
}
\author{
Pierre Riviere and Camille Vindras
}
