% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spatial.R
\name{model_spatial}
\alias{model_spatial}
\title{Run spatial row and column model}
\usage{
model_spatial(data, variable)
}
\arguments{
\item{data}{The data frame on which the model is run. 
It should come from \code{\link{format_data_PPBstats.data_agro}}}

\item{variable}{variable to analyse}
}
\value{
The function returns a list with two elements :
\itemize{
 \item info : a list with variable
 \item model : 
  \itemize{
   \item model: the output from SpATS function
   \item summary: summary of the model
   \item df_residuals: effective degree of freedom of the residuals
   \item MSerror: Mean Square error
  }
}
}
\description{
\code{model_spatial} runs spatial row and column model based on the SpATS package
}
\details{
The model is run with the SpATS function of package SpATS. 
See ?SpATS for more information.

More information can be found in the book \href{https://priviere.github.io/PPBstats_book/family-1.html#spatial-analysis}{here}.
}
\seealso{
\itemize{
\item \code{\link{check_model}}
\item \code{\link{check_model.fit_model_spatial}}
}
}
\author{
Pierre Riviere
}
