% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{mutlivariate}
\alias{mutlivariate}
\title{Run multivariate analysis with function from FactoMineR}
\usage{
mutlivariate(data, vec_variables, FUN)
}
\arguments{
\item{data}{The data frame. It should come from \code{\link{format_data_PPBstats.data_agro}}}

\item{vec_variables}{vector of variables to use in the analysis}

\item{FUN}{function to apply coming from FactoMineR, for example PCA}
}
\value{
The function returns the FactoMineR object corresponding to FUN
}
\description{
\code{mutlivariate} runs multivariate analysis with function from FactoMineR
}
\details{
This function is useful to settle the right data set and quali.sup

More information can be found in the book : https://priviere.github.io/PPBstats_book/family-5.html
}
\author{
Pierre Riviere
}
