% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_groups.R
\name{parameter_groups}
\alias{parameter_groups}
\title{Get groups of parameters based on multivariate analysis}
\usage{
parameter_groups(list_out_check_model, parameter)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{analyse.outputs}}

\item{parameter}{The parameter on which to get the groups}

\item{nb.clust}{Number of cluster to get. -1 to get automatic clustering. See details.}
}
\value{
The function returns a data frame with for each parameter, the median for each variable and the cluster to which it belongs
}
\description{
\code{get.parameter.groups}
}
\details{
The function run a PCA on a matrix containing the value of each parameter for each variable.

Clusters are done with the \code{FactoMineR::HCPC} function.
You can tone the number of clusters with the argument \code{nb.clust}, which is the same as in \code{HCPC}
Type ?\code{HCPC} for more informations.

More details in the vignette (type vignette ("PPBstats")).
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{FWH}}, \code{\link{analyse.outputs}}, \code{\link{get.ggplot}}
}

