% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_groups.R
\name{parameter_groups}
\alias{parameter_groups}
\title{Get groups of parameters based on multivariate analysis}
\usage{
parameter_groups(list_out_check_model, parameter)
}
\arguments{
\item{list_out_check_model}{A list whose elements are output from \code{\link{check_model}}}

\item{parameter}{The parameter on which to get the groups}
}
\value{
The function returns a list of two elements:
\itemize{
\item obj.pca : the PCA object from FactoMineR::PCA
\item clust, a list of two elements:
 \itemize{
 \item res.hcpc : the HCPC object from FactoMineR::HCPC
 \item clust : the dataframe with cluster assigned to each individual
 }
}
}
\description{
\code{parameter_groups} gets groups of parameters based on multivariate analysis
}
\details{
The function run a PCA on a matrix containing the value of each parameter for each variable.

Clusters are done based on HCPC method with FactoMineR package as explained here \url{http://www.sthda.com/english/wiki/hcpc-hierarchical-clustering-on-principal-components-hybrid-approach-2-2-unsupervised-machine-learning}
}
\seealso{
\itemize{
 \item \code{\link{check_model}},
 \item \code{\link{parameter_groups_GxE}},
 \item \code{\link{parameter_groups_model_bh_GxE}},
 \item \code{\link{plot.PPBstats}}
}
}
\author{
Pierre Riviere
}
