% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plan_experiment.R
\name{plan_experiment}
\alias{plan_experiment}
\title{Provides experimental plan for several situations}
\usage{
plan_experiment(expe.type, entries, controls, nb.controls.per.block, nb.blocks,
  nb.cols, return.format = "standard")
}
\arguments{
\item{expe.type}{The type of experiment to settle "satellite-farm", "regional-farm", "row-column", "fully-repicated", "IBD".}

\item{entries}{Vector with the names of the entries}

\item{controls}{Vector with names of controls}

\item{nb.controls.per.block}{Number of controls per blocks.}

\item{nb.blocks}{Number of blocks}

\item{nb.cols}{Number of columns in the design. The number of rows is computed automaticaly}

\item{return.format}{"standard" (entries, block, X, Y) or "shinemas" for SHiNeMaS reproduction template file}
}
\value{
The function returns a list with
\itemize{
 \item A data frame according to return.format
 \item A picture of the experimental plan
 }
}
\description{
\code{plan_experiment} provides experimental plan for several situations
}
\details{
Note that expe.type = "row-column" is particular case of expe.type = "regional-farm" where the number of controls must be at least the number of columns or rows.
The algoritms are the following:
\itemize{
 \item "satellite-farm"
 \enumerate{
	 \item randomize the entries
  \item get the data frame with one block, two columns and one control in each column
	 \item randomize the columns
 }
 \item "regional-farm"
 \enumerate{
	 \item randomize the entries
  \item get the data frame with blocks, columns and controls in each block
  \item arrange at least one control per row
  \item for each row, put control in different column
  \item randomize rows and columns
  \item check that controls do not touch each other
  \item check number of controls in col and row and send a warning message if control are missing in rows or columns
 }
 \item "row-column"
 \enumerate{
	 \item randomize the entries
  \item get the data frame with blocks, columns and controls in each block
  \item arrange at least one control per row
  \item for each row, put control in different column
  \item randomize rows and columns
  \item check that controls do not touch each other
  \item check number of controls in col and row and send a stop message if control are missing in rows or columns
 }
 \item "fully-repicated"
 \enumerate{
	 \item randomize the entries
  \item get the data frame with blocks, columns and entries in each block
  \item arrange randomization so that no entries can be in the same column between blocks
 }
 \item "IBD"
the randomization is based on the ibd function in the ibd package. See ?ibd for more information
}
}
\author{
Pierre Riviere
}

