% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plan_experiment.R
\name{plan_experiment}
\alias{plan_experiment}
\title{Provides experimental plan for several situations}
\usage{
plan_experiment(expe.type, nb.entries, nb.controls.per.block, nb.blocks,
  nb.cols)
}
\arguments{
\item{expe.type}{The type of experiment to settle "satellite-farm", "regional-farm", "row-column", "fully-repicated", "IBD".}

\item{nb.entries}{Number of entries}

\item{nb.controls.per.block}{Number of controls per blocks.}

\item{nb.blocks}{Number of blocks}

\item{nb.cols}{Number of columns in the design. The number of rows is computed automaticaly}
}
\value{
The function returns a list with
\itemize{
 \item A data frame with entries, block, X, Y
 \item A picture of the experimental plan
 }
}
\description{
\code{plan_experiment} provides experimental plan for several situations
}
\details{
Note that expe.type = "row-column" is particular case of expe.type = "regional-farm" where the number of controls must be at least the number of columns or rows.
}
\author{
Pierre Riviere
}

