% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_model_home_away.R
\name{plot.check_model_home_away}
\alias{plot.check_model_home_away}
\title{Get ggplot to visualize output from \code{\link{check_model.fit_model_home_away}}}
\usage{
\method{plot}{check_model_home_away}(x)
}
\arguments{
\item{x}{Output from \code{\link{check_model.fit_model_home_away}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
\itemize{
 \item residuals
 \itemize{
  \item histogram : histogram with the distribution of the residuals
  \item qqplot
  }
 \item variability_repartition : pie with repartition of SumSq for each factor
 \item variance_intra_location : repartition of the residuals for each location (see Details for more information)
 }
}
\description{
\code{plot.check_model_home_away} returns ggplot to visualize outputs from \code{\link{check_model.fit_model_home_away}}
}
\details{
S3 method.

variance_intra_location represents the repartition of the residuals for each location.
   This has to been seen with caution:
   \itemize{
    \item If location have no intra-location variance (i.e. pure line or hybrides) then the distribution of each location represent only the micro-environmental variation.
    \item If location have intra-location variance (i.e. population such as landraces for example) then the distribution of each location represent the micro-environmental variation plus the intra-location variance.
    With the hypothesis than the micro-environmental variation is equaly distributed on all the individuals (i.e. all the plants), the distribution of each location represent the intra-location variance.
   }

See example in the book \href{https://priviere.github.io/PPBstats_book/family-4.html#family-4}{here}
}
\seealso{
\code{\link{check_model.fit_model_home_away}}
}
\author{
Pierre Riviere and Baptiste Rouger
}
