% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_mean_comparisons_GxE.R
\name{plot.mean_comparisons_model_GxE}
\alias{plot.mean_comparisons_model_GxE}
\title{Get ggplot from mean_comparisons_model_GxE}
\usage{
\method{plot}{mean_comparisons_model_GxE}(x, nb_parameters_per_plot = 8)
}
\arguments{
\item{out_x}{outputs from \code{\link{x}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_x} returns ggplot from \code{\link{x}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}},
\item \code{\link{x}}
}
}

