% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mean_comparisons_model_spatial.R
\name{plot.mean_comparisons_model_spatial}
\alias{plot.mean_comparisons_model_spatial}
\title{Get ggplot from mean_comparisons_model_spatial}
\usage{
\method{plot}{mean_comparisons_model_spatial}(x, nb_parameters_per_plot = 8)
}
\arguments{
\item{out_x}{outputs from \code{\link{x}}}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_x} returns ggplot from \code{\link{x}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}}, 
\item \code{\link{x}}
}
}
