% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_parameter_groups.R
\name{plot.parameter_groups}
\alias{plot.parameter_groups}
\title{Get ggplot from parameter_groups}
\usage{
\method{plot}{parameter_groups}(x)
}
\arguments{
\item{x}{outputs from \code{\link{parameter_groups}} function}
}
\value{
See \code{\link{get_ggplot}}
}
\description{
\code{ggplot_parameter_groups} returns ggplot from \code{\link{parameter_groups}}
}
\details{
See \code{\link{get_ggplot}}
}
\seealso{
\itemize{
\item \code{\link{get_ggplot}}, 
\item \code{\link{parameter_groups}}
}
}
