% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{pmap}
\alias{pmap}
\title{map background for plot.data_agro, plot.data_network}
\usage{
pmap(net, format, labels_on, labels_size, zoom)
}
\arguments{
\item{net}{network object or data frame}

\item{format}{network format}

\item{labels_on}{where display label}

\item{labels_size}{label size}

\item{zoom}{zoom of the map}
}
\description{
map background for plot.data_agro, plot.data_network
}
