% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{package}
\name{PPBstats}
\alias{PPBstats}
\alias{PPBstats-package}
\title{PPBstats}
\description{
The R package \code{PPBstats} analyzes unbalanced trials from decentralized participatory plant breeding programs.
}
\details{
How to use \code{PPBstats} is explained in the vignette: type vignette("PPBstats").
}
\author{
Pierre Riviere and Olivier David
}

