% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPBstats-package.R
\docType{package}
\name{PPBstats}
\alias{PPBstats}
\alias{PPBstats-package}
\title{PPBstats}
\description{
\code{PPBstats} is an R package for statistical analysis of balanced and unbalanced trials in decentralized participatory plant breeding programmes
}
\details{
How to use \code{PPBstats} is explained in the vignette: type vignette("PPBstats").
}
\author{
Pierre Riviere, Gaelle Van Franck, Olivier David and Facundo Muñoz
}
